% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_alignments.R
\name{alignments_comparison_multiple}
\alias{alignments_comparison_multiple}
\title{Does something F}
\usage{
alignments_comparison_multiple(sim_pars = c(0.2, 0.15, 2, 0.15),
  max_sims = 10, chain_length = 1e+06, sample_interval = 1000,
  sequence_length = 1000, cond = 1, age = 10, tips_interval = c(0, Inf),
  mutation_rate = 1/age)
}
\arguments{
\item{sim_pars}{something}

\item{max_sims}{something}

\item{chain_length}{something}

\item{sample_interval}{something}

\item{sequence_length}{something}

\item{cond}{Set 1 if you want to condition on stem or crown age
and non-extinction of the phylogeny. Set 0 otherwise.}

\item{age}{The age of the tree.}

\item{tips_interval}{Sets tips boundaries constrain on simulated dataset.}

\item{mutation_rate}{something}
}
\description{
Does something F
}
