% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_alignments.R
\name{alignments_comparison_single}
\alias{alignments_comparison_single}
\title{Does something E}
\usage{
alignments_comparison_single(sim_phylo, chain_length = 1e+07,
  sample_interval = 1000, sequence_length = 1000, max_repetitions = 10,
  mutation_rate = 1/max(abs(ape::branching.times(sim_phylo))))
}
\arguments{
\item{sim_phylo}{something}

\item{chain_length}{something}

\item{sample_interval}{something}

\item{sequence_length}{something}

\item{max_repetitions}{something}

\item{mutation_rate}{something}
}
\description{
Does something E
}
