% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_alignments.R
\name{BD.infer.lambda.from.mutations}
\alias{BD.infer.lambda.from.mutations}
\title{Does something D}
\usage{
BD.infer.lambda.from.mutations(n_subs, mbd_lambda, mu, age,
  init_n_lineages = 2, sequence_length = 1000, mutation_rate = 1/age,
  n_steps = 40)
}
\arguments{
\item{mu}{something}

\item{age}{The age of the tree.}

\item{sequence_length}{something}

\item{mutation_rate}{something}
}
\description{
Does something D
}
