% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_utils.R
\name{bd_phylo2L}
\alias{bd_phylo2L}
\title{Convert bd phylo to L table}
\usage{
bd_phylo2L(phylo)
}
\arguments{
\item{phylo}{a phylogeny}
}
\value{
the L table
}
\description{
Convert bd phylo to L table. Don't use for mbd.
}
\author{
Xu Liang
}
