% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_n_taxas.R
\name{calc_n_taxas}
\alias{calc_n_taxas}
\title{Calculate the number of taxa that trees will have when
simulated from the set of MBD parameters.}
\usage{
calc_n_taxas(mbd_paramses)
}
\arguments{
\item{mbd_paramses}{data frame with one MBD parameter set per row.
If a simulation with one MBD parameter set has 10 replicates,
this data frame will hold 10 rows, each with different seeds}
}
\value{
a numeric vector of whole numbers,
  with as much values as the \code{mbd_paramses},
  containing the number of taxa each MBD parameter set will create
}
\description{
Calculate the number of taxa that trees will have when
simulated from the set of MBD parameters.
}
\author{
Richel J.C. Bilderbeek
}
