% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mbd_params.R
\name{check_mbd_params}
\alias{check_mbd_params}
\title{Create mbd parameters}
\usage{
check_mbd_params(mbd_params)
}
\arguments{
\item{mbd_params}{MBD parameter set}
}
\description{
Create mbd parameters
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
