% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mbd_paramses.R
\name{check_mbd_paramses}
\alias{check_mbd_paramses}
\title{Check all elements in the list of \code{mbd_params}.}
\usage{
check_mbd_paramses(mbd_paramses)
}
\arguments{
\item{mbd_paramses}{data frame with one MBD parameter set per row.
If a simulation with one MBD parameter set has 10 replicates,
this data frame will hold 10 rows, each with different seeds}
}
\description{
For example, each element should have a unique MBD parameter setting
}
\examples{
library(testthat)

mbd_paramses <- create_mbd_paramses()
expect_silent(check_mbd_paramses(mbd_paramses))
}
\author{
Richel J.C. Bilderbeek
}
