% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_misc_params.R
\name{check_misc_params}
\alias{check_misc_params}
\title{Check the misc parameters}
\usage{
check_misc_params(misc_params)
}
\arguments{
\item{misc_params}{additional parameters for razzo. They contain
tree_filename to store the original given tree and mbd_sim_rng_seed for
when an mbd tree is simulated}
}
\description{
Check the misc parameters
}
\author{
Giovanni Laudanno
}
