% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_project_folder_name}
\alias{check_project_folder_name}
\title{Check if \code{project_folder_name} ends with \code{razzo_project}
and is the name of an existing folder.
Will throw if not.
Else will do nothing.}
\usage{
check_project_folder_name(project_folder_name)
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
nothing
}
\description{
Check if \code{project_folder_name} ends with \code{razzo_project}
and is the name of an existing folder.
Will throw if not.
Else will do nothing.
}
\author{
Richel J.C. Bilderbeek
}
