% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_razzo_experiments.R
\name{check_razzo_experiments}
\alias{check_razzo_experiments}
\title{Check razzo experiments}
\usage{
check_razzo_experiments(experiments)
}
\arguments{
\item{experiments}{provides information on how the experiment has to be set
up. To create it use \code{\link[pirouette]{create_experiment}}}
}
\description{
Check razzo experiments
}
