% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_razzo_params.R
\name{check_razzo_params}
\alias{check_razzo_params}
\title{Check if the \code{razzo_params} are valid razzo parameters.}
\usage{
check_razzo_params(razzo_params)
}
\arguments{
\item{razzo_params}{a parameter set for one \code{razzo} experiment,
as created by \link{create_razzo_params}
and used by \link{run_razzo}}
}
\description{
Will \link{stop} if the \code{razzo_params} are invalid,
else will do nothing
}
\author{
Richel J.C. Bilderbeek
}
