% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params_razzo.R
\name{check_razzo_params_names}
\alias{check_razzo_params_names}
\title{Check if the \code{razzo_params} is a list with elements
that have the required names to be valid razzo parameters.}
\usage{
check_razzo_params_names(razzo_params)
}
\arguments{
\item{razzo_params}{a parameter set for one \code{razzo} experiment,
as created by \link{create_params_razzo}
and used by \link{run_razzo}}
}
\description{
Will \link{stop} if not
}
\author{
Richèl J.C. Bilderbeek
}
