% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{clean_path}
\alias{clean_path}
\title{Clean up a path by removing double folder seperators}
\usage{
clean_path(filename)
}
\arguments{
\item{filename}{name of a file}
}
\description{
Clean up a path by removing double folder seperators
}
\examples{
clean_path("a//b")
}
\author{
Richèl J.C. Bilderbeek
}
