% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_esses.R
\name{collect_esses}
\alias{collect_esses}
\title{Collect esses}
\usage{
collect_esses(project_folder_name = get_razzo_path("razzo_project"))
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
a dataframe with parameters and esses
}
\description{
Collect esses
}
\author{
Giovanni Laudanno
}
