% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_esses_from_log_filenames.R
\name{collect_esses_from_log_filenames}
\alias{collect_esses_from_log_filenames}
\title{Collect ESSes from log filenames}
\usage{
collect_esses_from_log_filenames(parameters, files_log)
}
\description{
Collect ESSes from log filenames
}
