% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_marg_liks.R
\name{collect_evidences}
\alias{collect_evidences}
\alias{collect_marg_liks}
\title{Collect marginal loglikelihoods}
\usage{
collect_evidences(project_folder_name = get_razzo_path("razzo_project"))
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
a dataframe with parameters and marginal likelihoods
}
\description{
Collect marginal loglikelihoods
}
\author{
Giovanni Laudanno
}
