% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_mbd_params.R
\name{collect_mbd_params}
\alias{collect_mbd_params}
\title{Collect mbd parameters}
\usage{
collect_mbd_params(project_folder_name = getwd())
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
a dataframe with parameters and folder
}
\description{
Collect mbd parameters
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
