% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_n_mb_species.R
\name{collect_n_mb_species}
\alias{collect_n_mb_species}
\title{Collect the number of multiple-born (MB) species}
\usage{
collect_n_mb_species(project_folder_name = getwd())
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
a dataframe with folder and number of multiple-born (MB) species
}
\description{
Collect the number of multiple-born (MB) species
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
