% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_n_mutations.R
\name{collect_n_mutations}
\alias{collect_n_mutations}
\title{Collect number of mutations}
\usage{
collect_n_mutations(project_folder_name = getwd())
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
a dataframe with folder and number of mutations
}
\description{
Collect number of mutations
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
