% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_n_taxa.R
\name{collect_n_taxa}
\alias{collect_n_taxa}
\title{Collect number of taxa}
\usage{
collect_n_taxa(project_folder_name = getwd())
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
a dataframe with folder and taxa number for each phylogeny
}
\description{
Collect number of taxa
}
\examples{
library(testthat)
library(raztr)

df <- collect_n_taxa(
  project_folder_name = get_raztr_path("razzo_project")
)

expect_true("folder" \%in\% names(df))
expect_true("n_taxa" \%in\% names(df))
expect_true(is.character(df$folder))
expect_true(is.numeric(df$n_taxa))
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
