% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_nltt_stats.R
\name{collect_nltt_stats}
\alias{collect_nltt_stats}
\title{Collect nltt statistics}
\usage{
collect_nltt_stats(project_folder_name = getwd())
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
a dataframe with parameters and nltt statistics
}
\description{
Collect nltt statistics
}
\examples{
library(raztr)

# Collect the nLTT statistics from the data supplied with this package
project_folder_name <- get_raztr_path("razzo_project")
collect_nltt_stats(
  project_folder_name = project_folder_name
)
}
\author{
Giovanni Laudanno, Richèl J.C. Bilderbeek
}
