% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_results.R
\name{collect_results}
\alias{collect_results}
\title{Collect all results}
\usage{
collect_results(project_folder_name = get_razzo_path("razzo_project"),
  include_all_nltt = FALSE)
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}

\item{include_all_nltt}{TRUE if you want ALL the nltt data (which might be
a lot)}
}
\value{
a dataframe that summarizes all the results
}
\description{
Collect all results
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
