% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bd_alignment_file.R
\name{create_bd_alignment_file}
\alias{create_bd_alignment_file}
\title{Create the twin BD alignment from a parameters file in the same folder}
\usage{
create_bd_alignment_file(parameters_filename)
}
\arguments{
\item{parameters_filename}{full path to a 'parameters.csv' file}
}
\value{
nothing
}
\description{
Create the twin BD alignment
  from a parameters file in the same folder
as that parameters file.
For example, for a path '/my_folder/parameters.csv', this function creates:
'/my_folder/bd.fasta'.
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
