% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bd_nltts_file.R
\name{create_bd_nltts_file}
\alias{create_bd_nltts_file}
\title{Create the NLTT statistics distribution files from a posterior file.
Assumes for a posterior file named '1x.trees'
  that there is a 'true' tree file called '1x.tree'}
\usage{
create_bd_nltts_file(parameters_filename)
}
\arguments{
\item{parameters_filename}{full path to a 'parameters.csv' file}
}
\value{
nltt statistics
}
\description{
Create the NLTT statistics distribution files from a posterior file.
Assumes for a posterior file named '1x.trees'
  that there is a 'true' tree file called '1x.tree'
}
\author{
Richel J.C. Bilderbeek
}
