% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bd_posterior_files.R
\name{create_bd_posterior_files}
\alias{create_bd_posterior_files}
\title{Create the inference files from a FASTA file.
For example, for a FASTA file named '/my_folder/mbd.fasta', this
function will create:
\itemize{
  \item '/my_folder/bd.trees' BEAST2 posterior trees
  \item '/my_folder/bd.log' BEAST2 parameter estimates
  \item '/my_folder/bd_marg_lik.csv' BEAST2 marginal likelihood estimate
}
Assumes, for a FASTA file named '/my_folder/bd.fasta', this
  that there is a parameters file named '/my_folder/parameters.csv'}
\usage{
create_bd_posterior_files(parameters_filename)
}
\arguments{
\item{parameters_filename}{full path to a 'parameters.csv' file}
}
\value{
names of the files created
}
\description{
Create the inference files from a FASTA file.
For example, for a FASTA file named '/my_folder/mbd.fasta', this
function will create:
\itemize{
  \item '/my_folder/bd.trees' BEAST2 posterior trees
  \item '/my_folder/bd.log' BEAST2 parameter estimates
  \item '/my_folder/bd_marg_lik.csv' BEAST2 marginal likelihood estimate
}
Assumes, for a FASTA file named '/my_folder/bd.fasta', this
  that there is a parameters file named '/my_folder/parameters.csv'
}
\author{
Richel J.C. Bilderbeek
}
