% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bd_tree_file.R
\name{create_bd_tree_files}
\alias{create_bd_tree_files}
\title{Create an BD twin tree from an MBD tree
and save it as a file}
\usage{
create_bd_tree_files(parameters_filename)
}
\arguments{
\item{parameters_filename}{full path to a 'parameters.csv' file}
}
\value{
The twin BD tree obtained from the corresponding MBD tree.
  It will create a file with name \code{bd.tree}
  in the folder relative to the chosen parameters
}
\description{
Create an BD twin tree from an MBD tree
and save it as a file
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
