% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fig_0_a.R
\name{create_fig_0_a}
\alias{create_fig_0_a}
\title{Create old figure 1: the nltt plot}
\usage{
create_fig_0_a(project_folder_name)
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
the nltt plot
}
\description{
Create old figure 1: the nltt plot
}
\author{
Giovanni Laudanno
}
