% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fig_0_b.R
\name{create_fig_0_b}
\alias{create_fig_0_b}
\title{Create old figure 1: the marginal logliks plot}
\usage{
create_fig_0_b(project_folder_name)
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
the marginal logliks plot
}
\description{
Create old figure 1: the marginal logliks plot
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
