% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fig_1.R
\name{create_fig_1}
\alias{create_fig_1}
\title{Create figure 1}
\usage{
create_fig_1(project_folder_name = get_razzo_path("razzo_project"))
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}

\item{model_type}{It can be either "gen" (stands for "generative") or "best"
(stands for "best candidate").}
}
\value{
Figure 1 as a ggplot2 plot
}
\description{
Create figure 1
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
