% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fig_1_file.R
\name{create_fig_1_file}
\alias{create_fig_1_file}
\title{Save nllt figure}
\usage{
create_fig_1_file(project_folder_name)
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
the file name of the plot
}
\description{
Save nllt figure
}
\author{
Giovanni Laudanno
}
