% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fig_2.R
\name{create_fig_2}
\alias{create_fig_2}
\title{Create figure 2}
\usage{
create_fig_2(project_folder_name)
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
Figure 2 as a ggplot2 plot
}
\description{
Create figure 2
}
\author{
Richel J.C. Bilderbeek
}
