% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fig_2_file.R
\name{create_fig_2_file}
\alias{create_fig_2_file}
\title{Create figure 2 to visualize how}
\usage{
create_fig_2_file(project_folder_name = getwd())
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
the name of the file the plot is saved to
}
\description{
Create figure 2 to visualize how
}
\author{
Richel J.C. Bilderbeek
}
