% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_all_data_file.R
\name{create_file_all_data}
\alias{create_file_all_data}
\alias{create_all_data_file}
\title{Collect all data
in one \code{.csv} file called \code{results/all_data.csv}}
\usage{
create_file_all_data(project_folder_name = getwd(), new_run = FALSE)
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}

\item{new_run}{If true will collect all data from scratch}
}
\value{
the name of the file created
}
\description{
Collect all data
in one \code{.csv} file called \code{results/all_data.csv}
}
\author{
Giovanni Laudanno
}
