% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_file_mbd_params.R
\name{create_file_mbd_params}
\alias{create_file_mbd_params}
\alias{create_mbd_params_file}
\title{Collect all MBD parameters of a \code{razzo} run and put these
in one \code{.csv} file called \code{results/mbd_params.csv}}
\usage{

  create_file_mbd_params(project_folder_name = get_razzo_path("razzo_project"))
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
the name of the file created
}
\description{
Collect all MBD parameters of a \code{razzo} run and put these
in one \code{.csv} file called \code{results/mbd_params.csv}
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
