% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_n_mutations_file.R
\name{create_file_n_mutations}
\alias{create_file_n_mutations}
\alias{create_n_mutations_file}
\title{Creates file containing the number of mutations}
\usage{
create_file_n_mutations(project_folder_name = getwd())
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
the name of the saved file
}
\description{
Creates file containing the number of mutations
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
