% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_file_n_taxa.R
\name{create_file_n_taxa}
\alias{create_file_n_taxa}
\alias{create_n_taxa_file}
\title{Creates file containing the number of taxa}
\usage{
create_file_n_taxa(project_folder_name = getwd())
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
the name of the saved file
}
\description{
Creates file containing the number of taxa
}
\author{
Richèl J.C. Bilderbeek
}
