% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_file_nltt_stats.R
\name{create_file_nltt_stats}
\alias{create_file_nltt_stats}
\alias{create_nltt_stats_file}
\title{Creates file containing nltt stats}
\usage{
create_file_nltt_stats(project_folder_name = getwd())
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
the name of the saved file
}
\description{
Creates file containing nltt stats
}
\author{
Giovanni Laudanno
}
