% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_file_run_times.R
\name{create_file_run_times}
\alias{create_file_run_times}
\alias{create_run_times_file}
\title{Creates file containing the run times}
\usage{
create_file_run_times(project_folder_name = getwd())
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
the name of the saved file
}
\description{
Creates file containing the run times
}
\author{
Richèl J.C. Bilderbeek
}
