% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameters_files.R
\name{create_full_parameters_files}
\alias{create_full_parameters_files}
\title{Create the parameter files as used in the experiment
  \code{project_folder_name/data/[settings]/seed/[models]}}
\usage{
create_full_parameters_files(project_folder_name = getwd())
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
Create folders for each parameter setting
  and saves each setting in a file within the corresponding folder.
}
\description{
Create the parameter files as used in the experiment
  \code{project_folder_name/data/[settings]/seed/[models]}
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
