% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_mbd_alignment_file.R
\name{create_mbd_alignment_file}
\alias{create_mbd_alignment_file}
\title{Create an MBD alignment from a parameters file in the same folder}
\usage{
create_mbd_alignment_file(parameters_filename)
}
\arguments{
\item{parameters_filename}{full path to a 'parameters.csv' file}
}
\value{
full path to the created file
}
\description{
Create an MBD alignment
  from a parameters file in the same folder
as that parameters file.
For example, for a path '/my_folder/parameters.csv', this function creates:
'/my_folder/mbd.fasta'.
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
