% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mbd_marg_lik_file.R
\name{create_mbd_marg_lik_file}
\alias{create_mbd_marg_lik_file}
\title{Create a file that stores the results of a marginal likelihood estimation
of an MBD alignment}
\usage{
create_mbd_marg_lik_file(parameters_filename)
}
\arguments{
\item{parameters_filename}{full path to a 'parameters.csv' file}
}
\value{
name of the created file
}
\description{
Create a file that stores the results of a marginal likelihood estimation
of an MBD alignment
}
\author{
Richel J.C. Bilderbeek
}
