% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mbd_params.R
\name{create_mbd_params}
\alias{create_mbd_params}
\title{Creates a list of mbd parameters,}
\usage{
create_mbd_params(lambda, mu, nu, q, cond, crown_age, seed)
}
\arguments{
\item{lambda}{per-lineage speciation rate}

\item{mu}{per-species extinction rate}

\item{nu}{the rate at which a multiple-birth specation is triggered}

\item{q}{something}

\item{crown_age}{The crown age of the tree.}

\item{seed}{a random number generator seed}
}
\value{
a list of parameters
}
\description{
Creates a list of mbd parameters,
}
\author{
Giovanni Laudanno
}
