% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mbd_params_interval.R
\name{create_mbd_params_interval}
\alias{create_mbd_params_interval}
\title{Create the parameter interval for mbd setting}
\usage{
create_mbd_params_interval(lambda = 0.2, mu = 0.15, nu = c(1, 1.5,
  2), q = c(0.1, 0.15, 0.2), seed = seq(from = 1, to = 2, by = 1),
  crown_age = 15, cond = 1)
}
\arguments{
\item{lambda}{per-lineage speciation rate. See \code{\link[mbd]{mbd_sim}}}

\item{mu}{per-species extinction rate. See \code{\link[mbd]{mbd_sim}}}

\item{nu}{the rate at which a multiple-birth specation is triggered.
See \code{\link[mbd]{mbd_sim}}}

\item{q}{per-species speciation probability in case of occurrance of
a multiple event. See \code{\link[mbd]{mbd_sim}}}

\item{seed}{a random number generator seed}

\item{crown_age}{The crown age of the tree.}

\item{cond}{conditioning as specified in \code{\link[mbd]{mbd_sim}}}
}
\description{
Create the parameter interval for mbd setting
}
\author{
Giovanni Laudanno
}
