% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mbd_params_table.R
\name{create_mbd_params_table}
\alias{create_mbd_params_table}
\title{Create a table of which each row holds a unique MBD parameter combination.#' the parameter interval for mbd setting}
\usage{
create_mbd_params_table(lambda = 0.2, mu = c(0, 0.15), nu = c(1, 1.5,
  2, 2.5), q = c(0.1, 0.15, 0.2), n_replicates = 2, crown_age = 15,
  cond = 1)
}
\arguments{
\item{lambda}{per-lineage speciation rate. See \code{\link[mbd]{mbd_sim}}}

\item{mu}{per-species extinction rate. See \code{\link[mbd]{mbd_sim}}}

\item{nu}{the rate at which a multiple-birth specation is triggered.
See \code{\link[mbd]{mbd_sim}}}

\item{q}{per-species speciation probability in case of occurrance of
a multiple event. See \code{\link[mbd]{mbd_sim}}}

\item{n_replicates}{number of replicates}

\item{crown_age}{The crown age of the tree.}

\item{cond}{conditioning as specified in \code{\link[mbd]{mbd_sim}}}
}
\description{
Create a table of which each row holds a unique MBD parameter combination.#' the parameter interval for mbd setting
}
\examples{
library(testthat)

df <- create_mbd_params_table()
expect_true(is.data.frame(df))
expect_equal(nrow(unique(df)), nrow(df))
expect_true("lambda" \%in\% names(df))
expect_true("mu" \%in\% names(df))
expect_true("nu" \%in\% names(df))
expect_true("q" \%in\% names(df))
expect_true("crown_age" \%in\% names(df))
expect_true("seed" \%in\% names(df))
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
