% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mbd_paramses.R
\name{create_mbd_paramses}
\alias{create_mbd_paramses}
\title{Create a list of MBD parameters}
\usage{
create_mbd_paramses()
}
\value{
a list of \code{mbd_params}
}
\description{
Each element has a unique MBD parameter combination
}
\examples{
library(testthat)

df <- create_mbd_paramses()
}
\author{
Richel J.C. Bilderbeek
}
