% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mbd_tree_events.R
\name{create_mbd_tree_events}
\alias{create_mbd_tree_events}
\title{Create an MBD tree from the razzo parameters}
\usage{
create_mbd_tree_events(parameters, nu_events)
}
\arguments{
\item{parameters}{the razzo parameters}

\item{nu_events}{the number of nu-triggered events that have to be
present in the simulated tree}
}
\value{
a list with elements that resulted from a
simulation under the MBD process:
\itemize{
  \item mbd_tree the reconstructed tree of class \code{phylo}
  \item l_matrix the L matrix
}
}
\description{
This works like create_mbd_tree but instead of specifying
the parameter nu, the user has to specify the number of nu-events in
the simulated process
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
