% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mbd_tree_file.R
\name{create_mbd_tree_files}
\alias{create_mbd_tree_files}
\title{Create an MBD tree from the razzo parameters
and save it as two files: \code{mbd.tree} and \code{mbd_l_matrix.csv}}
\usage{
create_mbd_tree_files(parameters_filename)
}
\arguments{
\item{parameters_filename}{full path to a 'parameters.csv' file}
}
\value{
The full path to the file named \code{mbd.tree}.
  This file contains a phylogeny simulated under the MBD process.
  The file will be saved in the same folder as the parameter file.
}
\description{
Create an MBD tree from the razzo parameters
and save it as two files: \code{mbd.tree} and \code{mbd_l_matrix.csv}
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
