% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_misc_params.R
\name{create_misc_params}
\alias{create_misc_params}
\title{Create the misc parameters}
\usage{
create_misc_params(tree_filename, mbd_sim_rng_seed)
}
\arguments{
\item{tree_filename}{name of the phylogeny file}

\item{mbd_sim_rng_seed}{rng seed to simulate a mbd tree}
}
\description{
Create the misc parameters
}
\author{
Giovanni Laudanno
}
