% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_n_mb_species_file.R
\name{create_n_mb_species_file}
\alias{create_n_mb_species_file}
\title{Create the dataframe with number of multiple-born species
and save it as a file}
\usage{
create_n_mb_species_file(project_folder_name = getwd())
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
the filename
}
\description{
Create the dataframe with number of multiple-born species
and save it as a file
}
\author{
Richèl J.C. Bilderbeek
}
