% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nltts.R
\name{create_nltts}
\alias{create_nltts}
\title{Create the NLTT statistics distribution files from a posterior file.
Assumes for a posterior file named '1x.trees'
  that there is a 'true' tree file called '1x.tree'}
\usage{
create_nltts(tree, posterior_trees)
}
\arguments{
\item{tree}{an ultrametric phylogenetic tree of class \code{phylo}}

\item{posterior_trees}{phylogenetic trees in a BEAST2 posterior,
of class \code{multiphylo}}
}
\value{
nltt statistics
}
\description{
Create the NLTT statistics distribution files from a posterior file.
Assumes for a posterior file named '1x.trees'
  that there is a 'true' tree file called '1x.tree'
}
\author{
Richel J.C. Bilderbeek
}
