% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameters_files.R
\name{create_parameters_files}
\alias{create_parameters_files}
\title{Create all parameter files  in
  \code{project_folder_name/data/[settings]/seed}}
\usage{
create_parameters_files(project_folder_name = getwd())
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
Create folders for each parameter setting
  and saves each setting in a file within the corresponding folder.
}
\description{
Create all parameter files  in
  \code{project_folder_name/data/[settings]/seed}
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
