% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_params_mbd.R
\name{create_params_mbd}
\alias{create_params_mbd}
\alias{create_mbd_params}
\title{Create mbd parameters}
\usage{
create_params_mbd(lambda, mu, nu, q, cond = 1, crown_age = 10, seed = NA)
}
\arguments{
\item{lambda}{per-lineage speciation rate. See \code{\link[mbd]{mbd_sim}}}

\item{mu}{per-species extinction rate. See \code{\link[mbd]{mbd_sim}}}

\item{nu}{the rate at which a multiple-birth specation is triggered.
See \code{\link[mbd]{mbd_sim}}}

\item{q}{per-species speciation probability in case of occurrance of
a multiple event. See \code{\link[mbd]{mbd_sim}}}

\item{cond}{conditioning as specified in \code{\link[mbd]{mbd_sim}}}

\item{crown_age}{The crown age of the tree.}

\item{seed}{a random number generator seed}
}
\description{
Create mbd parameters
}
\author{
Richèl J.C. Bilderbeek, Giovanni Laudanno
}
