% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_params_misc.R
\name{create_params_misc}
\alias{create_params_misc}
\alias{create_misc_params}
\title{Create the parameters for one experiment.
Run one point of the experiment}
\usage{
create_params_misc(tree_filename = peregrine::get_pff_tempfile(fileext =
  ".tree"))
}
\arguments{
\item{tree_filename}{name of the phylogeny file}
}
\description{
Create the parameters for one experiment.
Run one point of the experiment
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
