% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_posterior.R
\name{create_posterior}
\alias{create_posterior}
\title{Create the posterior from parameters and an alignment.}
\usage{
create_posterior(parameters, alignment)
}
\arguments{
\item{parameters}{the razzo parameters}

\item{alignment}{a DNA alignment}
}
\value{
the posterior, as a list of these items:
\itemize{
  \item trees the posterior trees, as a \code{multiPhylo}
  \item estimates the posterior parameter estimates, as a \code{data.frame}
  \item \code{ns} the Nested Sampling information
}
}
\description{
Create the posterior from parameters and an alignment.
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
