% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_razzo_alignment_params.R
\name{create_razzo_alignment_params}
\alias{create_razzo_alignment_params}
\title{Create a razzo alignment parameter set that matches
the razzo file naming convenstions and article}
\usage{
create_razzo_alignment_params(folder_name)
}
\description{
Create a razzo alignment parameter set that matches
the razzo file naming convenstions and article
}
\author{
Richel J.C. Bilderbeek
}
