% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_razzo_beast2_options.R
\name{create_razzo_beast2_options}
\alias{create_razzo_beast2_options}
\title{Create BEAST2 options that follow the razzo naming conventions
and article}
\usage{
create_razzo_beast2_options(model_type,
  folder_name = peregrine::get_pff_tempfile(), rng_seed = 1)
}
\arguments{
\item{rng_seed}{the RNG seed used in BEAST2 inference}
}
\description{
Create BEAST2 options that follow the razzo naming conventions
and article
}
