% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_razzo_cand_experiments.R
\name{create_razzo_cand_experiments}
\alias{create_razzo_cand_experiments}
\title{Create the candidate experiments}
\usage{
create_razzo_cand_experiments(gen_experiment)
}
\arguments{
\item{gen_experiment}{the generative experiment}
}
\description{
Create the candidate experiments
}
