% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_razzo_experiments.R
\name{create_razzo_experiments}
\alias{create_razzo_experiments}
\title{Creates a \code{pir_params} with the \code{razzo} setup
and naming scheme}
\usage{
create_razzo_experiments(has_candidates = FALSE,
  folder_name = peregrine::get_pff_tempfile(), rng_seed = 1)
}
\arguments{
\item{has_candidates}{if there are candidate experiments yes/no}

\item{folder_name}{name of the main folder}

\item{rng_seed}{the RNG seed used in inference}
}
\description{
Creates a \code{pir_params} with the \code{razzo} setup
and naming scheme
}
