% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_razzo_gen_experiment.R
\name{create_razzo_gen_experiment}
\alias{create_razzo_gen_experiment}
\title{Create a generative experiment that follows the razzo
naming conventions and article}
\usage{
create_razzo_gen_experiment(
  folder_name = peregrine::get_pff_tempfile(),
  rng_seed = 1
)
}
\arguments{
\item{folder_name}{name of the main folder}

\item{rng_seed}{RNG seed used in inference}
}
\description{
Create a generative experiment that follows the razzo
naming conventions and article
}
\author{
Richèl J.C. Bilderbeek
}
