% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_razzo_mrca_prior.R
\name{create_razzo_mrca_prior}
\alias{create_razzo_mrca_prior}
\title{Create the MRCA prior as used in the article}
\usage{
create_razzo_mrca_prior()
}
\description{
Create the MRCA prior as used in the article
}
