% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_razzo_params.R
\name{create_razzo_params}
\alias{create_razzo_params}
\title{Create the parameters for one experiment.
Run one point of the experiment}
\usage{
create_razzo_params(mbd_params,
  twinning_params = pirouette::create_twinning_params(),
  alignment_params = pirouette::create_alignment_params(root_sequence =
  "acgt", mutation_rate = 0.1),
  model_select_params = list(pirouette::create_gen_model_select_param(alignment_params),
  pirouette::create_best_model_select_param()),
  inference_params = pirouette::create_inference_params(),
  error_measure_params = pirouette::create_error_measure_params())
}
\description{
Create the parameters for one experiment.
Run one point of the experiment
}
\author{
Richel J.C. Bilderbeek
}
