% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_razzo_pir_params.R
\name{create_razzo_pir_params}
\alias{create_razzo_pir_params}
\title{Creates a \code{pir_params} with the \code{razzo} setup
and naming scheme}
\usage{
create_razzo_pir_params(has_candidates = FALSE, has_twinning = TRUE,
  folder_name = peregrine::get_pff_tempfile(), rng_seed = 1,
  error_measure_params = pirouette::create_error_measure_params())
}
\arguments{
\item{has_candidates}{if there are candidate experiments yes/no}

\item{has_twinning}{do use twinning yes/no}

\item{folder_name}{name of the main folder}

\item{rng_seed}{RNG seed for alignment simulation and inference}

\item{error_measure_params}{parameters to set how the error
between given tree and inferred trees in measure,
as can be created by \code{\link[pirouette]{create_error_measure_params}}}
}
\description{
Creates a \code{pir_params} with the \code{razzo} setup
and naming scheme
}
