% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_razzo_twinning_params.R
\name{create_razzo_twinning_params}
\alias{create_razzo_twinning_params}
\title{Create a \code{twinning_params} that follows the razzo naming
convention and article}
\usage{
create_razzo_twinning_params(folder_name)
}
\arguments{
\item{folder_name}{name of the main folder}
}
\description{
Create a \code{twinning_params} that follows the razzo naming
convention and article
}
\author{
Richel J.C. Bilderbeek
}
