% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_mbd_paramses.R
\name{create_test_mbd_paramses}
\alias{create_test_mbd_paramses}
\title{Create a set of testing MBD parameters}
\usage{
create_test_mbd_paramses()
}
\description{
Create a set of testing MBD parameters
}
\author{
Richel J.C. Bilderbeekc
}
