% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameters_files.R
\name{create_test_parameters_files}
\alias{create_test_parameters_files}
\title{Create all testing parameter files in
  \code{project_folder_name/data/[settings]/seed/[models]}}
\usage{
create_test_parameters_files(project_folder_name = getwd(),
  n_replicates = 2, root_sequence = "acgt", verbose = FALSE)
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}

\item{n_replicates}{number of replicates}

\item{verbose}{give more output}
}
\value{
Create folders for each parameter setting
  and saves each setting in a file within the corresponding folder.
}
\description{
Create all testing parameter files in
  \code{project_folder_name/data/[settings]/seed/[models]}
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
