% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_razzo_experiments.R
\name{create_test_razzo_experiments}
\alias{create_test_razzo_experiments}
\title{Create a set of testing experiments}
\usage{
create_test_razzo_experiments(has_candidates = TRUE)
}
\description{
Create a set of testing experiments
}
