% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_razzo_params.R
\name{create_test_razzo_params}
\alias{create_test_razzo_params}
\title{Create some testing parameters for one experiment.
Run one point of the experiment}
\usage{
create_test_razzo_params()
}
\description{
Create some testing parameters for one experiment.
Run one point of the experiment
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
