% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_razzo_paramses.R
\name{create_test_razzo_paramses}
\alias{create_test_razzo_paramses}
\title{Create a list of \code{razzo_params} to be used in testing}
\usage{

  create_test_razzo_paramses(project_folder_name = file.path(peregrine::get_pff_tempfile(),
  "razzo_project"))
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\description{
Create a list of \code{razzo_params} to be used in testing
}
