% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_razzo_filenames.R
\name{get_alignment_filename}
\alias{get_alignment_filename}
\title{Get alignment filename}
\usage{
get_alignment_filename(folder_name, tree_type)
}
\arguments{
\item{folder_name}{name of the main folder}

\item{tree_type}{can be 'true' or 'twin'}
}
\description{
Get alignment filename
}
