% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_paths.R
\name{get_data_paths}
\alias{get_data_paths}
\title{Get paths for all the settings}
\usage{
get_data_paths(project_folder_name)
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
the paths of all the parameter settings
}
\description{
Get paths for all the settings
}
\author{
Giovanni Laudanno
}
