% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_generative_model}
\alias{get_generative_model}
\title{Retrieve the generative model from the data}
\usage{
get_generative_model(seed_folder)
}
\value{
list of models
}
\description{
Retrieve the generative model from the data
}
\author{
Giovanni Laudanno
}
