% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_razzo_filenames.R
\name{get_output_trees_filenames}
\alias{get_output_trees_filenames}
\title{Get output trees filenames}
\usage{
get_output_trees_filenames(folder_name, model_type)
}
\arguments{
\item{folder_name}{name of the main folder}

\item{model_type}{the type of inference model, must be
\code{generative} or \code{candidate}}
}
\description{
Get output trees filenames
}
\author{
Giovanni Laudanno
}
