% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_path}
\alias{get_path}
\title{Get the full path of a file in the \code{inst/extdata} folder}
\usage{
get_path(filename)
}
\arguments{
\item{filename}{the file's name, without the path}
}
\value{
the full path of the filename, if and only if
  the file is present. Will stop otherwise.
}
\description{
Get the full path of a file in the \code{inst/extdata} folder
}
\examples{
  testit::assert(is.character(get_path("parameters.csv")))
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
