% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_razzo_crown_age.R
\name{get_razzo_crown_age}
\alias{get_razzo_crown_age}
\title{Get the crown age used in the article}
\usage{
get_razzo_crown_age()
}
\value{
the crown age
}
\description{
Get the crown age used in the article
}
\author{
Richel J.C. Bilderbeek
}
