% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_razzo_dna_alignment_length.R
\name{get_razzo_dna_alignment_length}
\alias{get_razzo_dna_alignment_length}
\title{Get the razzo DNA alignment length, as used in the article}
\usage{
get_razzo_dna_alignment_length()
}
\description{
Get the razzo DNA alignment length, as used in the article
}
