% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_razzo_lambdas.R
\name{get_razzo_lambdas}
\alias{get_razzo_lambdas}
\title{Get the \code{lambda}s used in the MBD process}
\usage{
get_razzo_lambdas()
}
\value{
a numeric vector of \code{lambda}s
}
\description{
Get the \code{lambda}s used in the MBD process
}
\author{
Richèl J.C. Bilderbeek
}
