% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_razzo_mcmc_chain_length.R
\name{get_razzo_mcmc_chain_length}
\alias{get_razzo_mcmc_chain_length}
\title{Get the MCMC chain length used in the article}
\usage{
get_razzo_mcmc_chain_length()
}
\value{
the MCMC chain length
}
\description{
Get the MCMC chain length used in the article
}
\author{
Richel J.C. Bilderbeek
}
