% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_razzo_mcmc_store_every.R
\name{get_razzo_mcmc_store_every}
\alias{get_razzo_mcmc_store_every}
\title{Get the MCMC sampling interval used in the article}
\usage{
get_razzo_mcmc_store_every()
}
\value{
the MCMC sampling interval
}
\description{
Get the MCMC sampling interval used in the article
}
\author{
Richel J.C. Bilderbeek
}
