% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_razzo_mus.R
\name{get_razzo_mus}
\alias{get_razzo_mus}
\title{Get the \code{mu}s used in the MBD process}
\usage{
get_razzo_mus()
}
\value{
a numeric vector of \code{mu}s
}
\description{
Get the \code{mu}s used in the MBD process
}
\author{
Richel J.C. Bilderbeek
}
