% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_razzo_nus.R
\name{get_razzo_nus}
\alias{get_razzo_nus}
\title{Get the \code{nu}s used in the MBD process}
\usage{
get_razzo_nus()
}
\value{
a numeric vector of \code{nu}s
}
\description{
Get the \code{nu}s used in the MBD process
}
\author{
Richèl J.C. Bilderbeek
}
