% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_razzo_qs.R
\name{get_razzo_qs}
\alias{get_razzo_qs}
\title{Get the \code{q}s used in the MBD process}
\usage{
get_razzo_qs()
}
\value{
a numeric vector of \code{q}s
}
\description{
Get the \code{q}s used in the MBD process
}
\author{
Richel J.C. Bilderbeek
}
