% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_results_path}
\alias{get_results_path}
\title{Get paths for the results}
\usage{
get_results_path(project_folder_name)
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
the paths of the results
}
\description{
Get paths for the results
}
\author{
Giovanni Laudanno
}
