% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_razzo_filenames.R
\name{get_seed_folder_name}
\alias{get_seed_folder_name}
\title{Get seed folder name}
\usage{
get_seed_folder_name(project_folder_name, mbd_params)
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}

\item{mbd_params}{MBD parameter set}
}
\description{
Get seed folder name
}
\author{
Giovanni Laudanno
}
