% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_razzo_filenames.R
\name{get_tracelog_filename}
\alias{get_tracelog_filename}
\title{Get tracelog filename}
\usage{
get_tracelog_filename(folder_name, model_type)
}
\arguments{
\item{folder_name}{name of the main folder}

\item{model_type}{the type of inference model, must be
\code{generative} or \code{candidate}}
}
\description{
Get tracelog filename
}
\author{
Giovanni Laudanno
}
