% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Returns a string to say hello to the world}
\usage{
hello()
}
\value{
a string to say hello to the world
}
\description{
Returns a string to say hello to the world
}
\author{
Hayley Wickham, modified by Richel J.C. Bilderbeek
}
