% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_one_bool.R
\name{is_one_bool}
\alias{is_one_bool}
\title{Check if the argument is one boolean}
\usage{
is_one_bool(x)
}
\arguments{
\item{x}{the argument to be tested to be boolean}
}
\description{
Check if the argument is one boolean
}
\examples{
library(testthat)
expect_true(is_one_bool(TRUE))
expect_true(is_one_bool(FALSE))
expect_false(is_one_bool(NULL))
expect_false(is_one_bool(NA))
expect_false(is_one_bool(c()))
expect_false(is_one_bool("nonsense"))
expect_false(is_one_bool(c(TRUE, FALSE)))
}
\author{
Richel J.C. Bilderbeek
}
