% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_ml_cluster.R
\name{mbd_ML_cluster}
\alias{mbd_ML_cluster}
\title{Maximization of the loglikelihood under
  a multiple birth-death diversification model, wrapped for cluster usage.}
\usage{
mbd_ML_cluster(s, initparsopt = c(0.6, 0.1, 1.3, 0.16), sim_data, cond, soc,
  tips_interval, brts, sim_pars, mbd_ML, missnumspec, minimum_multiple_births)
}
\arguments{
\item{s}{The number of the simulation you want to evaluate.}

\item{initparsopt}{The initial values of the parameters that must be optimized}

\item{sim_data}{something}

\item{cond}{Set 1 if you want to condition on stem or crown age
and non-extinction of the phylogeny. Set 0 otherwise.}

\item{soc}{Sets whether stem or crown age should be used (1 or 2).}

\item{tips_interval}{Sets tips boundaries constrain on simulated dataset.}

\item{brts}{something}

\item{sim_pars}{something}

\item{mbd_ML}{something}

\item{missnumspec}{something}

\item{minimum_multiple_births}{something}
}
\value{
The output is saved on the document "mbd_MLE.txt".
\itemize{
\item First column contains ML estimates for lambda.
\item Second column contains ML estimates for mu.
\item Third column contains ML estimates for lambda.
\item Fourth column contains maximum likelihood.
\item Fifth column contains the number of additional species coming from multiple births in the evaluated tree.
}
}
\description{
mbd_ML_cluster computes the maximum likelihood estimates of the parameters of a multiple birth-death diversification model. Differently from mbd_ML it needs only the number "s" of the simulations, making it suitable to run on a cluster. You will need two files to make it work: "general_settings","sim_data"; they are both generated by mbd_sim_dataset. This second version supports both sympatric and allopatric speciations.
}
\examples{
# @Giappo: does not work yet
#You will need two files to make it work: "general_settings","sim_data".
# mbd_ML_cluster(1)

}
\author{
Giovanni Laudanno
}
