% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbd_sim_dataset0.R
\name{mbd_sim_dataset0}
\alias{mbd_sim_dataset0}
\title{Creates a full simulated dataset of trees under the multiple birth death process}
\usage{
mbd_sim_dataset0(sim_pars = c(2.5, 0.1, 0.1), soc = 2, cond = 1,
  age = 10, max_sims = 1000, tips_interval = c(0, Inf), edge = Inf,
  minimum_multiple_births = 0)
}
\arguments{
\item{sim_pars}{something}

\item{soc}{Sets whether stem or crown age should be used (1 or 2).}

\item{cond}{Set 1 if you want to condition on stem or crown age
and non-extinction of the phylogeny. Set 0 otherwise.}

\item{age}{The age of the tree.}

\item{max_sims}{something}

\item{tips_interval}{Sets tips boundaries constrain on simulated dataset.}

\item{edge}{the program automatically detects the (estimated) average number of tips. "edge" defines the width of the spread around the mean.}

\item{minimum_multiple_births}{something}
}
\value{
The function returns a list of brts vectors, one for each of the max_sims simulations.
It also saves those in a file called "sim_data" and saves all the settings in a "general_settings" file.
N.B.: At each call of the function you overwrite the previous files.
}
\description{
mbd_sim_dataset0 produces a full dataset of max_sims simulated trees including multiple speciations at the same time.
}
\examples{
# @Giappo: does not work
# out <- mbd_sim_dataset0( pars=c(2.5,0.1,0.1), soc=2, age=10, cond=1, edge=Inf )

}
\author{
Giovanni Laudanno
}
