% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_taxa_and_mbness.R
\name{measure_taxa_and_mbness}
\alias{measure_taxa_and_mbness}
\title{Measure number of taxa and mb-ness for each parameter setting}
\usage{
measure_taxa_and_mbness(n_replicates = 10000)
}
\arguments{
\item{n_replicates}{number of replicates}
}
\value{
a dataframe with parameter settings, a sample of number of taxa and
 a sample of mb-ness (in percentage and absolute).
}
\description{
Measure number of taxa and mb-ness for each parameter setting
}
\author{
Giovanni Laudanno
}
