% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_alignments.R
\name{P_t1_t2}
\alias{P_t1_t2}
\title{Does something A}
\usage{
P_t1_t2(lambda, mu, t1, t2)
}
\arguments{
\item{lambda}{something}

\item{mu}{something}

\item{t1}{something}

\item{t2}{something}
}
\description{
Does something A
}
