% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmb_ml_cluster.R
\name{pmb_ML_cluster}
\alias{pmb_ML_cluster}
\title{Does something I}
\usage{
pmb_ML_cluster(s, initparsopt = c(0.5, 0, 1.7, 0.15))
}
\arguments{
\item{s}{something}

\item{initparsopt}{something}
}
\description{
Does something I
}
