% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmb_ml_cluster.R
\name{pmb_ML_cluster2}
\alias{pmb_ML_cluster2}
\title{Does something O}
\usage{
pmb_ML_cluster2(s, initparsopt = c(0.5, 0.15, 0.1), pmb_ML, sim_data, brts)
}
\arguments{
\item{s}{something}

\item{initparsopt}{something}

\item{pmb_ML}{something}

\item{sim_data}{something}

\item{brts}{something}
}
\description{
Does something O
}
