% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_n_extant_taxa.R
\name{predict_n_extant_taxa}
\alias{predict_n_extant_taxa}
\title{Calculate the expected number of extant taxa
at the present from the MBD parameters}
\usage{
predict_n_extant_taxa(pars, crown_age)
}
\arguments{
\item{pars}{numeric vector of length four, with values for
MBD parameters of lambda, mu, nu and q respectively}

\item{the}{crown age, time between the present and the
time there were only two lineages}
}
\value{
the expected number of extant taxa
}
\description{
Calculate the expected number of extant taxa
at the present from the MBD parameters
}
\author{
Richel J.C. Bilderbeek
}
