% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_n_extinct_taxa.R
\name{predict_n_extinct_taxa}
\alias{predict_n_extinct_taxa}
\title{Calculate the expected number of taxa
that went extinct before the present
from the MBD parameters}
\usage{
predict_n_extinct_taxa(pars, crown_age)
}
\arguments{
\item{pars}{numeric vector of length four, with values for
MBD parameters of lambda, mu, nu and q respectively}

\item{crown_age}{the crown age, time between the present and the
time there were only two lineages}
}
\value{
the expected number of extinct taxa
}
\description{
Calculate the expected number of taxa
that went extinct before the present
from the MBD parameters
}
\author{
Richel J.C. Bilderbeek
}
