% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_calc_mut_rates.R
\name{raz_calc_mut_rates}
\alias{raz_calc_mut_rates}
\title{Calculate the mutation rates}
\usage{
raz_calc_mut_rates(bd_tree, mbd_tree)
}
\arguments{
\item{bd_tree}{a phylogent of class \code{phylo}, created by a Birth Death
process}

\item{mbd_tree}{an MBD tree}
}
\value{
mutation rates
}
\description{
Analyze the twin trees to assign to each of them the right
mutation rate. In this a way it's possible to create alignments with the
same amount of information.
}
\author{
Giovanni Laudanno
}
