% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_collect_esses.R
\name{raz_collect_esses}
\alias{raz_collect_esses}
\title{Collect esses}
\usage{
raz_collect_esses(project_folder_name)
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
a dataframe with parameters and esses
}
\description{
Collect esses
}
\author{
Giovanni Laudanno
}
