% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_collect_mar_log_liks.R
\name{raz_collect_mar_log_liks}
\alias{raz_collect_mar_log_liks}
\title{Collect marginal loglikelihoods}
\usage{
raz_collect_mar_log_liks(filename)
}
\arguments{
\item{filename}{the file's name, without the path}
}
\value{
a dataframe with parameters and marginal likelihoods
}
\description{
Collect marginal loglikelihoods
}
\author{
Giovanni Laudanno
}
