% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_collect_marg_log_liks.R
\name{raz_collect_marg_liks}
\alias{raz_collect_marg_liks}
\title{Collect marginal loglikelihoods}
\usage{
raz_collect_marg_liks(project_folder_name)
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
a dataframe with parameters and marginal likelihoods
}
\description{
Collect marginal loglikelihoods
}
\author{
Giovanni Laudanno
}
