% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_collect_nltt_stats.R
\name{raz_collect_nltt_stats}
\alias{raz_collect_nltt_stats}
\title{Collect nltt statistics}
\usage{
raz_collect_nltt_stats(project_folder_name = getwd())
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
a dataframe with parameters and nltt statistics
}
\description{
Collect nltt statistics
}
\author{
Giovanni Laudanno
}
