% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_bd_alignment.R
\name{raz_create_bd_alignment}
\alias{raz_create_bd_alignment}
\title{Create the twin BD alignment from a parameters file in the same folder
as that parameters file.
For example, for a path '/my_folder/parameters.csv', this function creates:
'/my_folder/bd.fasta'.}
\usage{
raz_create_bd_alignment(parameters_filename)
}
\arguments{
\item{parameters_filename}{full path to a 'parameters.csv' file}
}
\value{
nothing
}
\description{
Create the twin BD alignment from a parameters file in the same folder
as that parameters file.
For example, for a path '/my_folder/parameters.csv', this function creates:
'/my_folder/bd.fasta'.
}
\author{
Richel J.C. Bilderbeek
}
