% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_bd_marg_lik_file.R
\name{raz_create_bd_marg_lik_file}
\alias{raz_create_bd_marg_lik_file}
\title{Create a file that stores the results of a marginal likelihood estimation
of an BD alignment}
\usage{
raz_create_bd_marg_lik_file(parameters_filename)
}
\value{
name of the created file
}
\description{
Create a file that stores the results of a marginal likelihood estimation
of an BD alignment
}
\author{
Richel J.C. Bilderbeek
}
