% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_bd_tree.R
\name{raz_create_bd_tree}
\alias{raz_create_bd_tree}
\title{Create an BD twin tree from an MBD tree
and save it as a file}
\usage{
raz_create_bd_tree(parameters, folder_name)
}
\arguments{
\item{parameters}{the razzo parameters}

\item{folder_name}{name of the folder}
}
\value{
A twin BD tree
}
\description{
Create an BD twin tree from an MBD tree
and save it as a file
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
