% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_esses_file.R
\name{raz_create_esses_file}
\alias{raz_create_esses_file}
\title{Create the dataframe with esses data
and save it as a file}
\usage{
raz_create_esses_file(project_folder_name)
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
the esses
}
\description{
Create the dataframe with esses data
and save it as a file
}
\author{
Giovanni Laudanno
}
