% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_filenames.R
\name{raz_create_filename.mbd_tree}
\alias{raz_create_filename.mbd_tree}
\title{Create filename: mbd_tree}
\usage{
raz_create_filename.mbd_tree(parameters, folder_name)
}
\arguments{
\item{parameters}{the razzo parameters}

\item{folder_name}{name of the main folder}
}
\value{
name of the mbd_tree file
}
\description{
Create filename: mbd_tree
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
