% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_filenames.R
\name{raz_create_filename_mbd_alignment}
\alias{raz_create_filename_mbd_alignment}
\title{Create filename: mbd_alignment}
\usage{
raz_create_filename_mbd_alignment(parameters, folder_name)
}
\arguments{
\item{parameters}{the razzo parameters}

\item{folder_name}{name of the main folder}
}
\value{
name of the mbd_alignment file
}
\description{
Create filename: mbd_alignment
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
