% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_inference_files.R
\name{raz_create_inference_files}
\alias{raz_create_inference_files}
\title{Create the inference files from a FASTA file:
\itemize{
  \item A a
  \item B aqwf
  \item C aewf
}
Assumes for a FASTA file named '1x.fasta'
  that there is a parameters file called '1.csv'}
\usage{
raz_create_inference_files(fasta_filename)
}
\value{
names of the files created
}
\description{
Create the inference files from a FASTA file:
\itemize{
  \item A a
  \item B aqwf
  \item C aewf
}
Assumes for a FASTA file named '1x.fasta'
  that there is a parameters file called '1.csv'
}
\author{
Richel J.C. Bilderbeek
}
