% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_input_files.R
\name{raz_create_input_files}
\alias{raz_create_input_files}
\title{Create the input files from a parameters file}
\usage{
raz_create_input_files(parameters_filename)
}
\value{
names of the files created
}
\description{
Create the input files from a parameters file
}
\author{
Richel J.C. Bilderbeek
}
