% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_mar_log_liks_file.R
\name{raz_create_mar_log_liks_file}
\alias{raz_create_mar_log_liks_file}
\title{Creates file containing marginal loglikelihoods}
\usage{
raz_create_mar_log_liks_file(project_folder_name)
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
the name of the saved file
}
\description{
Creates file containing marginal loglikelihoods
}
\author{
Giovanni Laudanno
}
