% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_mbd_tree_file.R
\name{raz_create_mbd_tree_file}
\alias{raz_create_mbd_tree_file}
\title{Create an MBD tree from the razzo parameters
and save it as a file}
\usage{
raz_create_mbd_tree_file(parameters, folder_name)
}
\arguments{
\item{parameters}{the razzo parameters}

\item{folder_name}{name of the main folder}
}
\value{
The tree simulated under the MBD process.
  It will create a file with name \code{mbd.tree}
  in the folder relative to the chosen parameters
}
\description{
Create an MBD tree from the razzo parameters
and save it as a file
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
