% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_nltt_file.R
\name{raz_create_nltt_file}
\alias{raz_create_nltt_file}
\title{Create the NLTT statistics distribution files from a posterior file.
Assumes for a posterior file named '1x.trees'
  that there is a 'true' tree file called '1x.tree'}
\usage{
raz_create_nltt_file(parameters, folder_name)
}
\arguments{
\item{parameters}{the razzo parameters}

\item{folder_name}{name of the main folder}
}
\value{
nltt statistics
}
\description{
Create the NLTT statistics distribution files from a posterior file.
Assumes for a posterior file named '1x.trees'
  that there is a 'true' tree file called '1x.tree'
}
\author{
Richel J.C. Bilderbeek
}
