% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_nltts.R
\name{raz_create_nltts}
\alias{raz_create_nltts}
\title{Create the NLTT statistics distribution files from a posterior file.
Assumes for a posterior file named '1x.trees'
  that there is a 'true' tree file called '1x.tree'}
\usage{
raz_create_nltts(tree, posterior_trees)
}
\value{
nltt statistics
}
\description{
Create the NLTT statistics distribution files from a posterior file.
Assumes for a posterior file named '1x.trees'
  that there is a 'true' tree file called '1x.tree'
}
\author{
Richel J.C. Bilderbeek
}
