% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_parameters.R
\name{raz_create_param_age}
\alias{raz_create_param_age}
\title{Create a parameter called age}
\usage{
raz_create_param_age(value = 0)
}
\value{
a parameter called age
}
\description{
Create a parameter called age
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
