% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_parameters.R
\name{raz_create_param_mbd_mutation_rate}
\alias{raz_create_param_mbd_mutation_rate}
\title{Create a parameter called mbd_mutation_rate}
\usage{
raz_create_param_mbd_mutation_rate(value = 0)
}
\arguments{
\item{value}{value of the parameter}
}
\value{
a parameter called mbd_mutation_rate
}
\description{
Create a parameter called mbd_mutation_rate
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
