% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_parameters.R
\name{raz_create_param_mu}
\alias{raz_create_param_mu}
\title{Create a parameter called mu}
\usage{
raz_create_param_mu(id = NA, value = 0)
}
\value{
a parameter called mu
}
\description{
Create a parameter called mu
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
