% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_parameters.R
\name{raz_create_param_nu}
\alias{raz_create_param_nu}
\title{Create a parameter called nu}
\usage{
raz_create_param_nu(value = 0)
}
\value{
a parameter called nu
}
\description{
Create a parameter called nu
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
