% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_parameters.R
\name{raz_create_param_q}
\alias{raz_create_param_q}
\title{Create a parameter called q}
\usage{
raz_create_param_q(value = 0)
}
\value{
a parameter called q
}
\description{
Create a parameter called q
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
