% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_parameters.R
\name{raz_create_param_sample_interval}
\alias{raz_create_param_sample_interval}
\title{Create a parameter called sample_interval}
\usage{
raz_create_param_sample_interval(value = 0)
}
\value{
a parameter called sample_interval
}
\description{
Create a parameter called sample_interval
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
