% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_parameters.R
\name{raz_create_param_seed}
\alias{raz_create_param_seed}
\title{Create a parameter called seed}
\usage{
raz_create_param_seed(value = 0)
}
\arguments{
\item{value}{value of the parameter}
}
\value{
a parameter called seed
}
\description{
Create a parameter called seed
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
