% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_parameters.R
\name{raz_create_param_sequence_length}
\alias{raz_create_param_sequence_length}
\title{Create a parameter called sequence_length}
\usage{
raz_create_param_sequence_length(value = 0)
}
\arguments{
\item{value}{value of the parameter}
}
\value{
a parameter called sequence_length
}
\description{
Create a parameter called sequence_length
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
