% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_parameters.R
\name{raz_create_param_soc}
\alias{raz_create_param_soc}
\title{Create a parameter called soc}
\usage{
raz_create_param_soc(value = 0)
}
\arguments{
\item{value}{value of the parameter}
}
\value{
a parameter called soc
}
\description{
Create a parameter called soc
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
