% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_parameters.R
\name{raz_create_param_sub_chain_length}
\alias{raz_create_param_sub_chain_length}
\title{Create a parameter called sub_chain_length}
\usage{
raz_create_param_sub_chain_length(value = 0)
}
\value{
a parameter called sub_chain_length
}
\description{
Create a parameter called sub_chain_length
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
