% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_parameters_files.R
\name{raz_create_parameters_files}
\alias{raz_create_parameters_files}
\title{For each parameter setting create a folder and a parameters file}
\usage{
raz_create_parameters_files(folder_name = getwd())
}
\arguments{
\item{folder_name}{name of the main folder}
}
\value{
Create folders for each parameter setting
  and saves each setting in a file within the corresponding folder.
}
\description{
For each parameter setting create a folder and a parameters file
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
