% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_tempfile.R
\name{raz_create_tempfile}
\alias{raz_create_tempfile}
\title{Copies a file from \code{inst/extdata} and puts it in
a temporary folder.
Will stop if the file is absent in \code{inst/extdata}.}
\usage{
raz_create_tempfile(filename)
}
\arguments{
\item{filename}{name of the file in \code{inst/extdata}}
}
\value{
the name of the temporary file. Will also copy the
  \code{inst/extdata} file to that temporary folder.
}
\description{
Copies a file from \code{inst/extdata} and puts it in
a temporary folder.
Will stop if the file is absent in \code{inst/extdata}.
}
\examples{
  filename <- raz_create_tempfile("parameters.csv")
  testthat::expect_true(file.exists(filename))
}
\author{
Richel J.C. Bilderbeek
}
