% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_est_marg_lik.R
\name{raz_est_marg_lik}
\alias{raz_est_marg_lik}
\title{Estimates the marginal likelihood of a model
(as specified in \code{parameters})
for an alignment.}
\usage{
raz_est_marg_lik(parameters, alignment)
}
\arguments{
\item{parameters}{the razzo parameters}

\item{alignment}{a DNA alignment}
}
\value{
a list with the following elements:
\itemize{
  \item log_lik the log of the estimated marginal likelihood
  \item log_error the ?log of the error in that estimation
}
}
\description{
Estimates the marginal likelihood of a model
(as specified in \code{parameters})
for an alignment.
}
\author{
Richel J.C. Bilderbeek
}
