% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_filenames.R
\name{raz_filename.bd_alignment}
\alias{raz_filename.bd_alignment}
\title{Create filename: bd_alignment}
\usage{
raz_filename.bd_alignment(parameters, folder_name)
}
\arguments{
\item{parameters}{the razzo parameters}

\item{folder_name}{name of the main folder}
}
\value{
name of the bd_alignment file
}
\description{
Create filename: bd_alignment
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
