% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_utils.R
\name{raz_get_clock_models}
\alias{raz_get_clock_models}
\title{Clock models in Razzo}
\usage{
raz_get_clock_models()
}
\value{
the clock models
}
\description{
Clock models in Razzo
}
\author{
Giovanni Laudanno
}
