% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_utils.R
\name{raz_get_gen_models}
\alias{raz_get_gen_models}
\title{Generative models in Razzo}
\usage{
raz_get_gen_models()
}
\value{
the generative models
}
\description{
Generative models in Razzo
}
\author{
Giovanni Laudanno
}
