% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_get_parameters_path.R
\name{raz_get_parameters_path}
\alias{raz_get_parameters_path}
\title{Produces the right path from parameters}
\usage{
raz_get_parameters_path(parameters, folder_name)
}
\arguments{
\item{parameters}{the razzo parameters}

\item{folder_name}{name of the folder}
}
\value{
Path
}
\description{
Produces the right path from parameters
}
\author{
Giovanni Laudanno
}
