% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_get_settings_paths.R
\name{raz_get_settings_paths}
\alias{raz_get_settings_paths}
\title{Get paths for all the settings}
\usage{
raz_get_settings_paths(filename)
}
\arguments{
\item{filename}{the file's name, without the path}
}
\value{
the paths of all the parameter settings
}
\description{
Get paths for all the settings
}
\author{
Giovanni Laudanno
}
