% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_utils.R
\name{raz_get_test_filenames}
\alias{raz_get_test_filenames}
\title{Get standard parameters names for tests}
\usage{
raz_get_test_filenames()
}
\value{
nothing
}
\description{
Get standard parameters names for tests
}
\author{
Giovanni Laudanno
}
