% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_utils.R
\name{raz_load_test_params}
\alias{raz_load_test_params}
\title{Load standard parameters interval for tests}
\usage{
raz_load_test_params()
}
\value{
nothing
}
\description{
Load standard parameters interval to test
  the function "raz_create_parameters_files".
It loads the dataset in the data folder of the package.
}
\author{
Giovanni Laudanno
}
