% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_utils.R
\name{raz_make_tempdir}
\alias{raz_make_tempdir}
\title{Create a temporary folder}
\usage{
raz_make_tempdir(...)
}
\arguments{
\item{...}{options for \code{tempdir}}
}
\value{
the address of the temporary folder
}
\description{
The function removes all the pre-existing temporary folders
  and creates a new temporary folder.
Mainly used to run tests.
}
\author{
Giovanni Laudanno
}
