% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_open_parameters_file.R
\name{raz_open_parameters_file}
\alias{raz_open_parameters_file}
\title{Opens a parameter file and parses it}
\usage{
raz_open_parameters_file(filename)
}
\arguments{
\item{filename}{name of the parameter filename,
for example '/my_folder/1/parameters.csv'}
}
\value{
the razzo parameters
}
\description{
Opens a parameter file and parses it
}
\author{
Richel J.C. Bilderbeek
}
