% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_utils.R
\name{raz_save_standard_test_parameters}
\alias{raz_save_standard_test_parameters}
\title{Create standard parameters interval for tests}
\usage{
raz_save_standard_test_parameters()
}
\value{
nothing
}
\description{
Create standard parameters interval to test the function "raz_create_parameters_files".
It saves the dataset in the data folder of the package.
}
\author{
Giovanni Laudanno
}
