% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_utils.R
\name{raz_standard_parameters}
\alias{raz_standard_parameters}
\title{Create standard parameters for tests}
\usage{
raz_standard_parameters()
}
\value{
the parameters
}
\description{
Create standard parameters for tests
}
\author{
Giovanni Laudanno
}
