% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_utils.R
\name{raz_standard_parameters_interval}
\alias{raz_standard_parameters_interval}
\title{Create standard parameters interval for tests}
\usage{
raz_standard_parameters_interval()
}
\value{
the whole parameters interval
}
\description{
Create standard parameters interval to test the function "raz_create_parameters_files".
}
\author{
Giovanni Laudanno
}
