% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_cbs_exps.R
\name{remove_cbs_exps}
\alias{remove_cbs_exps}
\title{Remove all experiments that have a CBS tree prior}
\usage{
remove_cbs_exps(experiments)
}
\arguments{
\item{experiments}{a list of one or more pirouette experiments}
}
\value{
experiments that do not have a CBS tree prior
}
\description{
Remove all experiments that have a CBS tree prior
}
\author{
Richel J.C. Bilderbeek
}
