% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_razzo.R
\name{run_razzo}
\alias{run_razzo}
\title{Runs one \code{razzo} experiment}
\usage{
run_razzo(razzo_params)
}
\arguments{
\item{razzo_params}{a parameter set for one \code{razzo} experiment,
as created by \link{create_razzo_params}
and used by \link{run_razzo}}
}
\value{
nothing, but it does create a lot of files
}
\description{
Runs one \code{razzo} experiment
}
\author{
Richel J.C. Bilderbeek
export
}
