% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_razzo_paramses.R
\name{save_razzo_paramses}
\alias{save_razzo_paramses}
\title{Save the list of \code{razzo_params} to the exected locations}
\usage{

  save_razzo_paramses(razzo_paramses = create_razzo_paramses(project_folder_name
  = peregrine::get_pff_tempfile()))
}
\arguments{
\item{razzo_paramses}{a list of \code{razzo_params} for a
set of \code{razzo} experiment.
Use create \link{create_razzo_paramses} or
\link{create_test_razzo_paramses} to create such a list}
}
\value{
the paths to each created \code{parameters.RDa} file
}
\description{
A \code{razzo_params} already hold its desired location, which
if the folder where
\code{razzo_params$pir_params$alignment_params$fasta_filename}
will be stored. That foldername already follows the
\code{project_folder_name/data/[settings]/seed/[models]}
file storage location convention.
This function will create a \code{parameters.RDa}
in that folder.
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
