% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_alignments.R
\name{BD.infer.lambda.from.mutations}
\alias{BD.infer.lambda.from.mutations}
\title{Does something D}
\usage{
BD.infer.lambda.from.mutations(Nsubs, mbd.lambda, mu, age, N0 = 2,
  sequence_length = 1000, mutation_rate = 1/age, Nsteps = 40)
}
\arguments{
\item{Nsubs}{something}

\item{mbd.lambda}{something}

\item{mu}{something}

\item{age}{The age of the tree.}

\item{N0}{something}

\item{sequence_length}{something}

\item{mutation_rate}{something}

\item{Nsteps}{something}
}
\description{
Does something D
}
