% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_alignments.R
\name{BD.Nct}
\alias{BD.Nct}
\title{Does something B}
\usage{
BD.Nct(lambda, mu, t, N0 = 2, age)
}
\arguments{
\item{lambda}{something}

\item{mu}{something}

\item{t}{something}

\item{N0}{something}

\item{age}{The age of the tree.}
}
\description{
Does something B
}
