% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_alignments.R
\name{BD.Nmutations}
\alias{BD.Nmutations}
\title{Does something C}
\usage{
BD.Nmutations(lambda, mu, age, init_n_lineages = 2, sequence_length = 1000,
  mutation_rate = 1/age)
}
\arguments{
\item{lambda}{something}

\item{mu}{something}

\item{age}{The age of the tree.}

\item{sequence_length}{something}

\item{mutation_rate}{something}
}
\description{
Does something C
}
