% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_razzo_params_filenames.R
\name{check_razzo_params_filenames}
\alias{check_razzo_params_filenames}
\title{Check if all filenames in the \code{razzo_params}
follow the \code{razzo} naming convention}
\usage{
check_razzo_params_filenames(razzo_params)
}
\description{
Check if all filenames in the \code{razzo_params}
follow the \code{razzo} naming convention
}
