% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{clean_path}
\alias{clean_path}
\title{Clean up a path by removing double folder seperators}
\usage{
clean_path(filename)
}
\arguments{
\item{filename}{name of a file}
}
\description{
Clean up a path by removing double folder seperators
}
\examples{
 library(testthat)

 if (rappdirs::app_dir()$os != "win") {
   testthat::expect_equal(clean_path("a//b"), "a/b")
 }
 testthat::expect_error(clean_path(NULL))
 testthat::expect_error(clean_path(NA))
 testthat::expect_error(clean_path(Inf))
}
\author{
Richel J.C. Bilderbeek
}
