% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{clean_paths}
\alias{clean_paths}
\title{Clean up a paths by removing double folder seperators}
\usage{
clean_paths(filenames)
}
\arguments{
\item{filenames}{names of files}
}
\description{
Clean up a paths by removing double folder seperators
}
\examples{
library(testthat)
if (rappdirs::app_dir()$os != "win") {
  expect_equal(
    razzo::clean_paths(c("a//b")),
    c("a/b")
  )
  expect_equal(
    razzo::clean_paths(c("a//b", "c//d")),
    c("a/b", "c/d")
  )
}
expect_error(razzo::clean_paths(NULL))
expect_error(razzo::clean_paths(NA))
expect_error(razzo::clean_paths(Inf))
}
\author{
Richel J.C. Bilderbeek
}
