% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{clean_paths}
\alias{clean_paths}
\title{Clean up a paths by removing double folder seperators}
\usage{
clean_paths(filenames)
}
\arguments{
\item{filenames}{names of files}
}
\description{
Clean up a paths by removing double folder seperators
}
\examples{
clean_paths(c("a//b", "c//d"))
}
\author{
Richèl J.C. Bilderbeek
}
