% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_n_mb_events.R
\name{collect_n_mb_events}
\alias{collect_n_mb_events}
\title{Collect the number of multiple-birth (MB) events}
\usage{

  collect_n_mb_events(project_folder_name = get_razzo_path("razzo_project"))
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
a dataframe with folder and number of multiple-birth (MB) events
}
\description{
Collect the number of multiple-birth (MB) events
}
\author{
Richel J.C. Bilderbeek
}
