% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_n_taxa.R
\name{collect_n_taxa}
\alias{collect_n_taxa}
\title{Collect number of taxa}
\usage{
collect_n_taxa(project_folder_name = get_razzo_path("razzo_project"))
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
a dataframe with folder and taxa number for each phylogeny
}
\description{
Collect number of taxa
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
