% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_nltt_stats.R
\name{collect_nltt_stats}
\alias{collect_nltt_stats}
\title{Collect nltt statistics}
\usage{
collect_nltt_stats(project_folder_name = get_path("razzo_project"))
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
a dataframe with parameters and nltt statistics
}
\description{
Collect nltt statistics
}
\author{
Giovanni Laudanno
}
