% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_run_times.R
\name{collect_run_times}
\alias{collect_run_times}
\title{Collect the times simulations took to run}
\usage{
collect_run_times(project_folder_name = getwd())
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
a data frame with columns named 'filename', 'state' and 'cpu_time'
}
\description{
Collect the times simulations took to run
}
\author{
Richel J.C. Bilderbeek
}
