% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_fig_1.R
\name{create_fig_1}
\alias{create_fig_1}
\title{Create nllt figure}
\usage{
create_fig_1(project_folder_name)
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
the nltt plot
}
\description{
Create nllt figure
}
\author{
Giovanni Laudanno
}
