% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fig_1.R
\name{create_fig_1}
\alias{create_fig_1}
\title{Create figure 1}
\usage{
create_fig_1(project_folder_name)
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
Figure 1 as a ggplot2 plot
}
\description{
Create figure 1
}
\author{
Giovanni Laudanno
}
