% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameters_files.R
\name{create_full_parameters_files}
\alias{create_full_parameters_files}
\title{Create the parameter files according to the specified arguments
  \code{project_folder_name/data/[settings]/seed/[models]}}
\usage{

  create_full_parameters_files(project_folder_name = get_path("razzo_project"),
  mbd_params_interval = create_mbd_params_interval(lambda = 0.2, mu =
  0.15, nu = c(1, 1.5, 2), q = c(0.1, 0.15, 0.2), seed = seq(from = 1, to =
  2, by = 1), crown_age = 15, cond = 1),
  twinning_params = pirouette::create_twinning_params(),
  alignment_params = pirouette::create_alignment_params(root_sequence =
  "aaaaccccggggttt", mutation_rate =
  0.5/unique(mbd_params_interval$crown_age)),
  error_measure_params = pirouette::create_error_measure_params())
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}

\item{mbd_params_interval}{data frame with one MBD parameter set per row.
If a simulation with one MBD parameter set has 10 replicates,
this data frame will hold 10 rows, each with different seeds}

\item{twinning_params}{parameters for creating a twin tree,
as can be created by \code{\link[pirouette]{create_twinning_params}}}

\item{alignment_params}{parameters for creating an alignment,
as can be created by \code{\link[pirouette]{create_alignment_params}}}

\item{error_measure_params}{parameters to set how the error
between given tree and inferred trees in measure,
as can be created by \code{\link[pirouette]{create_error_measure_params}}}
}
\value{
Create folders for each parameter setting
  and saves each setting in a file within the corresponding folder.
}
\description{
Create the parameter files according to the specified arguments
  \code{project_folder_name/data/[settings]/seed/[models]}
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
