% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mbd_params_table.R
\name{create_mbd_params_table}
\alias{create_mbd_params_table}
\title{Create a table of which each row holds a unique MBD parameter combination.}
\usage{
create_mbd_params_table(
  lambdas = get_razzo_lambdas(),
  mus = get_razzo_mus(),
  nus = get_razzo_nus(),
  qs = get_razzo_qs(),
  cond = 1,
  crown_age = get_razzo_crown_age(),
  n_replicates = get_razzo_n_replicates()
)
}
\arguments{
\item{lambdas}{per-lineage speciation rates. See \code{\link[mbd]{mbd_sim}}}

\item{mus}{per-species extinction rates. See \code{\link[mbd]{mbd_sim}}}

\item{nus}{the rates at which a multiple-birth specation is triggered.
See \code{\link[mbd]{mbd_sim}}}

\item{qs}{per-species speciation probabilities in case of occurrance of
a multiple event. See \code{\link[mbd]{mbd_sim}}}

\item{cond}{conditioning as specified in \code{\link[mbd]{mbd_sim}}}

\item{crown_age}{The crown age of the tree.}

\item{n_replicates}{number of replicates}
}
\value{
a data frame with the following columns:
\itemize{
  \item lambda the speciation rate
  \item mu the extinction rate
  \item nu the co-occurring speciation event rate
  \item q the proportion of species that speciates at a co-occurring
    speciation event
  \item crown_age the crown age of the MBD tree
  \item cond the conditioning
  \item seed the RNG seed for creating an MBD tree
}
}
\description{
Create a table of which each row holds a unique MBD parameter combination.
}
\examples{
create_mbd_params_table()
}
\seealso{
use \link{create_mbd_paramses} to create a list
of \code{mbd_params}
}
\author{
Giovanni Laudanno, Richèl J.C. Bilderbeek
}
