% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mbd_paramses.R
\name{create_mbd_paramses}
\alias{create_mbd_paramses}
\title{Create a list of MBD parameters}
\usage{
create_mbd_paramses(n_replicates = get_razzo_n_replicates())
}
\arguments{
\item{n_replicates}{number of replicates}
}
\value{
a list of \code{mbd_params}
}
\description{
Each element has a unique MBD parameter combination
}
\examples{
create_mbd_paramses()
}
\seealso{
use \link{create_mbd_params} to create one
  MBD parameter combination
}
\author{
Richèl J.C. Bilderbeek
}
