% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_misc_params.R
\name{create_misc_params}
\alias{create_misc_params}
\title{Create the parameters for one experiment.
Run one point of the experiment}
\usage{
create_misc_params(tree_filename = get_pff_tempfile(fileext = ".tree"))
}
\arguments{
\item{tree_filename}{name of the phylogeny file}
}
\description{
Create the parameters for one experiment.
Run one point of the experiment
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
