% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_params_misc.R
\name{create_misc_params}
\alias{create_misc_params}
\title{Create the parameters for one experiment.
Run one point of the experiment}
\usage{
create_misc_params(folder_name = peregrine::get_pff_tempfile())
}
\arguments{
\item{folder_name}{name of the main folder}
}
\description{
Create the parameters for one experiment.
Run one point of the experiment
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
