% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_files_razzo_paramses.R
\name{create_parameters_files}
\alias{create_parameters_files}
\alias{create_files_razzo_paramses}
\title{Create all parameter files  in
  \code{project_folder_name/data/[settings]/seed/[models]}}
\usage{
create_parameters_files(project_folder_name = getwd(),
  experiment_type = "test")
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}

\item{experiment_type}{the type of experiment,
can be either \code{test} or \code{full}}
}
\value{
Create folders for each parameter setting
  and saves each setting in a file within the corresponding folder.
}
\description{
Create all parameter files  in
  \code{project_folder_name/data/[settings]/seed/[models]}
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
