% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_params_razzo.R
\name{create_params_razzo}
\alias{create_params_razzo}
\alias{create_razzo_params}
\title{Create one set of parameter for one \code{razzo} experiment.}
\usage{
create_params_razzo(mbd_params, pir_params, misc_params)
}
\arguments{
\item{mbd_params}{MBD parameter set}

\item{pir_params}{a parameter set for one \code{pirouette} run,
as created by \link[pirouette]{create_pir_params}.
Use \link{create_test_razzo_pir_params} to create
a \code{pir_params} that has the file naming scheme as \code{razzo}}

\item{misc_params}{additional parameters for razzo. They contain
tree_filename to store the original given tree and mbd_sim_rng_seed for
when an mbd tree is simulated}
}
\description{
All the files follow the expected naming scheme,
using the \link{dirname} of
\code{pir_params$alignment_params$fasta_filename} as
the folder where all files will be created
}
\author{
Richèl J.C. Bilderbeek, Giovanni Laudanno
}
