% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_razzo_alignment_params.R
\name{create_razzo_alignment_params}
\alias{create_razzo_alignment_params}
\title{Create a razzo alignment parameter set that matches
the razzo file naming convenstions and article}
\usage{
create_razzo_alignment_params(folder_name, rng_seed = 1)
}
\arguments{
\item{folder_name}{name of the main folder}

\item{rng_seed}{RNG seed for the alignment simulation}
}
\description{
Create a razzo alignment parameter set that matches
the razzo file naming convenstions and article
}
\author{
Richèl J.C. Bilderbeek
}
