% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_razzo_cand_experiments.R
\name{create_razzo_cand_experiments}
\alias{create_razzo_cand_experiments}
\title{Create the candidate experiments}
\usage{
create_razzo_cand_experiments(gen_experiment,
  folder_name = peregrine::get_pff_tempfile(), rng_seed = 1)
}
\arguments{
\item{gen_experiment}{the generative experiment}

\item{folder_name}{name of the main folder}
}
\value{
a list of experiments
}
\description{
It needs to know what the generative experiment is,
to exclude it from the created candidate experiments.
}
\author{
Richel J.C. Bilderbeek
}
