% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_razzo_gen_experiment.R
\name{create_razzo_gen_experiment}
\alias{create_razzo_gen_experiment}
\title{Create a generative experiment that follows the razzo
naming conventions and article}
\usage{
create_razzo_gen_experiment()
}
\description{
Create a generative experiment that follows the razzo
naming conventions and article
}
