% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_razzo_params.R
\name{create_razzo_params}
\alias{create_razzo_params}
\title{Create the parameters for one experiment.
Run one point of the experiment}
\usage{
create_razzo_params(mbd_params,
  twinning_params = pirouette::create_twinning_params(),
  alignment_params, gen_model_select_params, best_model_select_params,
  inference_params, sampling_method, misc_params)
}
\arguments{
\item{misc_params}{additional parameters for razzo. They contain
tree_filename to store the original given tree and mbd_sim_rng_seed for
when an mbd tree is simulated}
}
\description{
Create the parameters for one experiment.
Run one point of the experiment
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
