% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_razzo_params.R
\name{create_razzo_params}
\alias{create_razzo_params}
\title{Create the parameters for one experiment.}
\usage{
create_razzo_params(mbd_params, pir_params, misc_params)
}
\arguments{
\item{mbd_params}{MBD parameter set}

\item{pir_params}{a parameter set for one \code{pirouette} run,
as created by \link[pirouette]{create_pir_params}}

\item{misc_params}{additional parameters for razzo. They contain
tree_filename to store the original given tree and mbd_sim_rng_seed for
when an mbd tree is simulated}
}
\description{
Run one point of the experiment
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
