% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_razzo_params.R
\name{create_razzo_params}
\alias{create_razzo_params}
\title{Create the parameters for one experiment.
Run one point of the experiment}
\usage{
create_razzo_params(mbd_params,
  twinning_params = pirouette::create_twinning_params(),
  alignment_params, inference_params, model_select_params)
}
\description{
Create the parameters for one experiment.
Run one point of the experiment
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
