% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_razzo_pir_params.R
\name{create_razzo_pir_params}
\alias{create_razzo_pir_params}
\title{Creates a \code{pir_params} with the \code{razzo} setup
and naming scheme}
\usage{
create_razzo_pir_params(has_candidates = FALSE, has_twinning = TRUE,
  folder_name = peregrine::get_pff_tempfile())
}
\arguments{
\item{has_candidates}{if there are candidate experiments yes/no}

\item{has_twinning}{do use twinning yes/no}

\item{folder_name}{name of the main folder}
}
\description{
Creates a \code{pir_params} with the \code{razzo} setup
and naming scheme
}
