% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_razzo_twinning_params.R
\name{create_razzo_twinning_params}
\alias{create_razzo_twinning_params}
\title{Create a \code{twinning_params} that follows the razzo naming
convention and article}
\usage{
create_razzo_twinning_params(
  folder_name,
  rng_seed_twin_tree,
  rng_seed_twin_alignment
)
}
\arguments{
\item{rng_seed_twin_tree}{the random number generator seed as used in the
simulation of a twin tree}

\item{rng_seed_twin_alignment}{the random number generator seed
as used in the simulation of a twin alignment}
}
\description{
Create a \code{twinning_params} that follows the razzo naming
convention and article
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
