% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_params_mbd.R
\name{create_test_mbd_params}
\alias{create_test_mbd_params}
\alias{create_params_mbd_test}
\title{Create MBD parameters used in tests}
\usage{
create_test_mbd_params()
}
\description{
Create MBD parameters used in tests
}
\author{
Richel J.C. Bilderbeek
}
