% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameters_files.R
\name{create_test_parameters_files}
\alias{create_test_parameters_files}
\title{Create all testing parameter files in
  \code{project_folder_name/data/[settings]/seed/[models]}}
\usage{
create_test_parameters_files(project_folder_name = getwd(),
  n_replicates = 2, twinning_params = create_twinning_params(),
  alignment_params = create_alignment_params(root_sequence =
  "aaaaccccggggttt", mutation_rate = 0.5/15),
  inference_params = create_inference_params(mcmc =
  create_mcmc(chain_length = 2000, store_every = 1000), mrca_prior =
  create_mrca_prior(is_monophyletic = TRUE, mrca_distr =
  create_normal_distr(mean = 15, sigma = 0.01))),
  error_measure_params = create_error_measure_params())
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}

\item{n_replicates}{number of replicates}

\item{twinning_params}{parameters for creating a twin tree,
as can be created by \code{\link[pirouette]{create_twinning_params}}}

\item{alignment_params}{parameters for creating an alignment,
as can be created by \code{\link[pirouette]{create_alignment_params}}}

\item{inference_params}{the parameters used in all Bayesian inference,
as created by \link[pirouette]{create_inference_params}}

\item{error_measure_params}{parameters to set how the error
between given tree and inferred trees in measure,
as can be created by \code{\link[pirouette]{create_error_measure_params}}}
}
\value{
Create folders for each parameter setting
  and saves each setting in a file within the corresponding folder.
}
\description{
Create all testing parameter files in
  \code{project_folder_name/data/[settings]/seed/[models]}
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
