% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_pff_pir_params.R
\name{create_test_pff_pir_params}
\alias{create_test_pff_pir_params}
\title{Create a \code{pir_params} with PFFs}
\usage{
create_test_pff_pir_params()
}
\description{
Create a \code{pir_params} with PFFs
}
\author{
Richel J.C. Bilderbeek
}
