% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_razzo_params.R
\name{create_test_razzo_params}
\alias{create_test_razzo_params}
\title{Create some testing parameters for one experiment.
Run one point of the experiment}
\usage{
create_test_razzo_params(mbd_params = create_test_mbd_params(),
  pir_params = pirouette::create_test_pir_params(experiments =
  list(pirouette::create_test_gen_experiment(inference_model =
  beastier::create_inference_model(mrca_prior =
  beastier::create_mrca_prior(mrca_distr =
  beastier::create_normal_distr(mean = 15, sigma = 0.001), is_monophyletic
  = TRUE), mcmc = beautier::create_mcmc(chain_length = 2000, store_every =
  1000)))), twinning_params = pirouette::create_twinning_params()),
  misc_params = create_misc_params())
}
\arguments{
\item{mbd_params}{MBD parameter set}

\item{pir_params}{a parameter set for one \code{pirouette} run,
as created by \link[pirouette]{create_pir_params}}

\item{misc_params}{additional parameters for razzo. They contain
tree_filename to store the original given tree and mbd_sim_rng_seed for
when an mbd tree is simulated}
}
\description{
Create some testing parameters for one experiment.
Run one point of the experiment
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
