% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(abstol, account_name, alpha, alpha0, b,
  bd_tree_filename, brts, chain_length, changeloglikifnoconv, colormap,
  fasta_filename, folder_name, idparsfix, idparsopt, initparsopt,
  init_speciation_rate, init_extinction_rate, iterations, k, lambda, logs,
  lx, lx0, matrix, matrix_builder, max_iter, max_k, max_number_of_species,
  max_repetitions, max_sims, maxiter, mbd_tree, methode,
  minimum_multiple_births, missnumspec, mu, mutation_rate, nu, optimmethod,
  parameters, parameter_filename, parameters_filename, pars.transform,
  parsfix, precision, print_errors, q, reltol, res, s, sample_interval,
  sequence_length, sim_pars, sim_phylo, soc, t, t1, t2, time_interval, tol,
  transition_matrix, tree_filename, trees_filename, trparsfix, trparsopt,
  verbose, age, cond, pars, tips_interval)
}
\arguments{
\item{abstol}{something}

\item{account_name}{something}

\item{alpha}{something}

\item{alpha0}{something}

\item{b}{something}

\item{bd_tree_filename}{name of the file that stores a BD twin tree}

\item{brts}{something}

\item{chain_length}{something}

\item{changeloglikifnoconv}{something}

\item{colormap}{something}

\item{fasta_filename}{name of a FASTA file}

\item{folder_name}{name of the main folder}

\item{idparsfix}{something}

\item{idparsopt}{The ids of the parameters that must be optimized.
The ids are defined as follows:
\itemize{
  \item id == 1 corresponds to lambda (speciation rate)
  \item id == 2 corresponds to mu (extinction rate)
  \item id == 3 corresponds to nu (multiple speciation trigger rate)
  \item id == 4 corresponds to q (single-lineage speciation probability)
}}

\item{initparsopt}{something}

\item{init_speciation_rate}{a speciation rate}

\item{init_extinction_rate}{an extinction rate}

\item{iterations}{something}

\item{k}{something}

\item{lambda}{per-lineage speciation rate}

\item{logs}{something}

\item{lx}{It is the number of ODEs considered for the computation.}

\item{lx0}{something}

\item{matrix}{something}

\item{matrix_builder}{something}

\item{max_iter}{Sets the maximum number of iterations in the optimization}

\item{max_k}{something}

\item{max_number_of_species}{something}

\item{max_repetitions}{something}

\item{max_sims}{something}

\item{maxiter}{something}

\item{mbd_tree}{an MBD tree}

\item{methode}{something}

\item{minimum_multiple_births}{something}

\item{missnumspec}{something}

\item{mu}{something}

\item{mutation_rate}{something}

\item{nu}{something}

\item{optimmethod}{something}

\item{parameters}{the razzo parameters}

\item{parameter_filename}{full path to a 'parameters.csv' file}

\item{parameters_filename}{full path to a 'parameters.csv' file}

\item{pars.transform}{something}

\item{parsfix}{The values of the parameters that should not be optimized.}

\item{precision}{something}

\item{print_errors}{something}

\item{q}{something}

\item{reltol}{something}

\item{res}{something}

\item{s}{something}

\item{sample_interval}{something}

\item{sequence_length}{something}

\item{sim_pars}{something}

\item{sim_phylo}{something}

\item{soc}{Sets whether stem or crown age should be used (1 or 2).}

\item{t}{something}

\item{t1}{something}

\item{t2}{something}

\item{time_interval}{something}

\item{tol}{something}

\item{transition_matrix}{something}

\item{tree_filename}{name of the phylogeny file}

\item{trees_filename}{name of the BEAST2 posterior phylogenies file}

\item{trparsfix}{something}

\item{trparsopt}{something}

\item{verbose}{give more output}

\item{age}{The age of the tree.}

\item{cond}{Set 1 if you want to condition on stem or crown age
and non-extinction of the phylogeny. Set 0 otherwise.}

\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is the multiple speciation trigger rate;
  \item pars[2] is the extinction rate;
  \item pars[3] is the single-lineage speciation probability.
}
or
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability.
}}

\item{tips_interval}{Sets tips boundaries constrain on simulated dataset.}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Documentation by Giovanni Laudanno,
  use of this function by Richel J.C. Bilderbeek
}
