% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(alignment, bd_mutation_rate, bd_tree, bd_tree_filename,
  chain_length, crown_age, fasta_filename, folder_name,
  init_speciation_rate, init_extinction_rate, lambda, mbd_l_matrix,
  mbd_mutation_rate, mbd_tree, mu, mutation_rate, nu, parameters,
  parameter_filename, parameters_filename, posterior_trees, q,
  sample_interval, seed, sequence_length, sim_pars, sim_phylo,
  sub_chain_length, tree, tree_filename, trees_filename, verbose)
}
\arguments{
\item{alignment}{a DNA alignment}

\item{bd_mutation_rate}{the mutation rate when creating an alignment
from a BD tree}

\item{bd_tree}{a phylogent of class \code{phylo}, created by a Birth Death
process}

\item{bd_tree_filename}{name of the file that stores a BD twin tree}

\item{chain_length}{something}

\item{crown_age}{The crown age of the tree.}

\item{fasta_filename}{name of a FASTA file}

\item{folder_name}{name of the main folder}

\item{init_speciation_rate}{a speciation rate}

\item{init_extinction_rate}{an extinction rate}

\item{lambda}{per-lineage speciation rate}

\item{mbd_l_matrix}{the L matrix of an MBD tree}

\item{mbd_mutation_rate}{the mutation rate when creating an alignment
from a MBD tree}

\item{mbd_tree}{an MBD tree}

\item{mu}{per-species extinction rate}

\item{mutation_rate}{something}

\item{nu}{the rate at which a multiple-birth specation is triggered}

\item{parameters}{the razzo parameters}

\item{parameter_filename}{full path to a 'parameters.csv' file}

\item{parameters_filename}{full path to a 'parameters.csv' file}

\item{posterior_trees}{phylogenetic trees in a BEAST2 posterior,
of class \code{multiphylo}}

\item{q}{something}

\item{sample_interval}{the interval at which the MCMC algorithm
makes a measurement}

\item{seed}{a random number generator seed}

\item{sequence_length}{the length of each DNA sequence in an alignment}

\item{sim_pars}{something}

\item{sim_phylo}{something}

\item{sub_chain_length}{length of the sub-chain used by the Nested Sampling
algorithm to estimate the marginal likelihood}

\item{tree}{an ultrametric phylogenetic tree of class \code{phylo}}

\item{tree_filename}{name of the phylogeny file}

\item{trees_filename}{name of the BEAST2 posterior phylogenies file}

\item{verbose}{give more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Documentation by Giovanni Laudanno,
  use of this function by Richel J.C. Bilderbeek
}
