% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_best_model}
\alias{get_best_model}
\title{Retrieve the best candidates from the data}
\usage{
get_best_model(seed_folder)
}
\value{
list of models
}
\description{
Retrieve the best candidates from the data
}
\author{
Giovanni Laudanno
}
