% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_paths.R
\name{get_data_paths}
\alias{get_data_paths}
\title{Get paths for all the settings}
\usage{
get_data_paths(project_folder_name)
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}
}
\value{
the paths of all the parameter settings
}
\description{
Get paths for all the settings
}
\examples{
  # Obtain the paths of all razzo testing folders
  all_paths <- get_data_paths(get_path("razzo_project"))

  # In each of these, there is a 'parameters.RDa' file
  parameter_files <- file.path(all_paths, "parameters.RDa")
  testthat::expect_true(all(file.exists(parameter_files)))
}
\author{
Giovanni Laudanno
}
