% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_functions.R
\name{get_data_paths}
\alias{get_data_paths}
\title{Get the folder paths of \code{razzo} experiments}
\usage{
get_data_paths(project_folder_name, full_names = TRUE)
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}

\item{full_names}{TRUE if you want to return the full path}
}
\value{
the folder paths of all \code{razzo} experiments
}
\description{
This function looks recursively for \code{parameters.RDa} files
and return the folder names of these.
}
\examples{
library(testthat)

# Obtain the paths of all razzo testing folders
all_paths <- get_data_paths(raztr::get_raztr_path("razzo_project"))

# In each of these, there is a 'parameters.RDa' file
parameter_files <- file.path(all_paths, "parameters.RDa")
expect_true(all(file.exists(parameter_files)))
}
\author{
Giovanni Laudanno, Richèl J.C. Bilderbeek
}
