% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gen_models}
\alias{get_gen_models}
\title{Get the names of the supported generative models}
\usage{
get_gen_models()
}
\value{
the generative models
}
\description{
Get the names of the supported generative models
}
\examples{
get_gen_models()
}
\author{
Giovanni Laudanno, Richèl J.C. Bilderbeek
}
