% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_razzo_mcmc.R
\name{get_razzo_mcmc}
\alias{get_razzo_mcmc}
\title{Get the MCMC chain used in the article}
\usage{
get_razzo_mcmc()
}
\value{
the MCMC chain
}
\description{
Get the MCMC chain used in the article
}
\author{
Richel J.C. Bilderbeek
}
