% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_collect_esses.R
\name{raz_collect_esses}
\alias{raz_collect_esses}
\title{Collect esses}
\usage{
raz_collect_esses(filename)
}
\arguments{
\item{filename}{the file's name, without the path}
}
\value{
a dataframe with parameters and esses
}
\description{
Collect esses
}
\author{
Giovanni Laudanno
}
