% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_bd_alignment.R
\name{raz_create_bd_alignment}
\alias{raz_create_bd_alignment}
\title{Create a BD alignment from a BD tree}
\usage{
raz_create_bd_alignment(parameters, bd_tree, mbd_tree)
}
\arguments{
\item{parameters}{the razzo parameters}

\item{bd_tree}{a phylogent of class \code{phylo}, created by a Birth Death
process}

\item{mbd_tree}{an MBD tree}
}
\value{
an alignment
}
\description{
Create the twin BD alignment
  from parameters and a BD tree
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
