% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_mbd_alignment.R
\name{raz_create_mbd_alignment}
\alias{raz_create_mbd_alignment}
\title{Create and MBD alignment from a parameters file in the same folder
as that parameters file.
For example, for a path '/my_folder/parameters.csv', this function creates:
'/my_folder/mbd.fasta'.}
\usage{
raz_create_mbd_alignment(parameters_filename)
}
\arguments{
\item{parameters_filename}{full path to a 'parameters.csv' file}
}
\value{
nothing
}
\description{
Create and MBD alignment from a parameters file in the same folder
as that parameters file.
For example, for a path '/my_folder/parameters.csv', this function creates:
'/my_folder/mbd.fasta'.
}
\author{
Richel J.C. Bilderbeek
}
