% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_mbd_alignment.R
\name{raz_create_mbd_alignment}
\alias{raz_create_mbd_alignment}
\title{Create an MBD alignment from a parameters file in the same folder}
\usage{
raz_create_mbd_alignment(parameters, folder_name)
}
\arguments{
\item{parameters}{the razzo parameters}

\item{folder_name}{name of the main folder}
}
\value{
nothing
}
\description{
Create an MBD alignment
  from a parameters file in the same folder
as that parameters file.
For example, for a path '/my_folder/parameters.csv', this function creates:
'/my_folder/mbd.fasta'.
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
