% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_mbd_tree.R
\name{raz_create_mbd_tree}
\alias{raz_create_mbd_tree}
\title{Create an MBD tree from the razzo parameters}
\usage{
raz_create_mbd_tree(parameters)
}
\arguments{
\item{parameters}{the razzo parameters}
}
\value{
The trees simulated under the MBD process
}
\description{
Create an MBD tree from the razzo parameters
}
\author{
Giovanni Laudanno and Richel J.C. Bilderbeek
}
