% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_mbd_tree.R
\name{raz_create_mbd_tree}
\alias{raz_create_mbd_tree}
\title{Create an MBD tree from the razzo parameters
and save it as a file}
\usage{
raz_create_mbd_tree(parameters, tree_filename)
}
\arguments{
\item{parameters}{the razzo parameters}

\item{tree_filename}{name of the phylogeny file}
}
\value{
nothing. Will create a file with name in \code{tree_filename}
}
\description{
Create an MBD tree from the razzo parameters
and save it as a file
}
\author{
Richel J.C. Bilderbeek
}
