% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_mbd_tree.R
\name{raz_create_mbd_tree}
\alias{raz_create_mbd_tree}
\title{Create an MBD tree from the razzo parameters}
\usage{
raz_create_mbd_tree(parameters)
}
\arguments{
\item{parameters}{the razzo parameters}
}
\value{
a list with elements that resulted from a
simulation under the MBD process:
\itemize{
  \item mbd_tree the reconstructed tree of class \code{phylo}
  \item l_matrix the L matrix
}
}
\description{
Create an MBD tree from the razzo parameters
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
