% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_parameters.R
\name{raz_create_param}
\alias{raz_create_param}
\title{General function to create a parameter.}
\usage{
raz_create_param(name, value, ...)
}
\value{
a parameter
}
\description{
General function to create a parameter.
}
\note{
Prefer using the
  named functions
  \code{\link{create_param_lambda}},
  \code{\link{create_param_mu}},
  \code{\link{create_param_nu}},
  \code{\link{create_param_q}},
  \code{\link{create_param_seed}},
  \code{\link{create_param_cond}},
  \code{\link{create_param_age}},
  \code{\link{create_param_soc}},
  \code{\link{create_param_sequence_length}},
  \code{\link{create_param_mbd_mutation_rate}}
  and \code{\link{create_param_sample_interval}}
}
\author{
Richel J.C. Bilderbeek
}
