% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_parameters.R
\name{raz_create_param_lambda}
\alias{raz_create_param_lambda}
\title{Create a parameter called lambda}
\usage{
raz_create_param_lambda(value = 0)
}
\value{
a parameter called lambda
}
\description{
Create a parameter called lambda
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
