% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_parameters_files.R
\name{raz_create_parameters_files}
\alias{raz_create_parameters_files}
\title{Create all parameters files}
\usage{
raz_create_parameters_files()
}
\value{
names of the files created
}
\description{
Create all parameters files
}
\author{
Richel J.C. Bilderbeek
}
