% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_params.R
\name{raz_create_params}
\alias{raz_create_params}
\title{Creates a list of parameters,
after checking the arguments for their validity}
\usage{
raz_create_params(lambda = 0.1, mu = 0.05, nu = 0.1, q = 0.1,
  seed = 42, crown_age = 15, sequence_length = 100,
  mbd_mutation_rate = 1/crown_age, bd_mutation_rate = 1/crown_age,
  sample_interval = 1000, chain_length = 2000,
  sub_chain_length = 500)
}
\arguments{
\item{lambda}{per-lineage speciation rate}

\item{mu}{per-species extinction rate}

\item{nu}{the rate at which a multiple-birth specation is triggered}

\item{q}{something}

\item{seed}{a random number generator seed}

\item{crown_age}{The crown age of the tree.}

\item{sequence_length}{the length of each DNA sequence in an alignment}

\item{mbd_mutation_rate}{the mutation rate when creating an alignment
from a MBD tree}

\item{bd_mutation_rate}{the mutation rate when creating an alignment
from a BD tree}

\item{sample_interval}{the interval at which the MCMC algorithm
makes a measurement}

\item{chain_length}{something}

\item{sub_chain_length}{length of the sub-chain used by the Nested Sampling
algorithm to estimate the marginal likelihood}
}
\value{
a list of parameters
}
\description{
Creates a list of parameters,
after checking the arguments for their validity
}
\author{
Richel J.C. Bilderbeek
}
