% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_create_params.R
\name{raz_create_params}
\alias{raz_create_params}
\title{Creates a list of parameters,
after checking the arguments for their validity}
\usage{
raz_create_params(lambda)
}
\arguments{
\item{lambda}{per-lineage speciation rate}
}
\value{
a list of parameters
}
\description{
Creates a list of parameters,
after checking the arguments for their validity
}
\author{
Richel J.C. Bilderbeek
}
