% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raz_open_parameters_file.R
\name{raz_open_parameters_file}
\alias{raz_open_parameters_file}
\title{Opens a parameter file and parses it}
\usage{
raz_open_parameters_file(parameters_filename)
}
\arguments{
\item{parameters_filename}{full path to a 'parameters.csv' file}
}
\value{
the razzo parameters
}
\description{
Opens a parameter file and parses it
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
