% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_cbs_exps.R
\name{remove_cbs_exps}
\alias{remove_cbs_exps}
\title{Remove all experiments that have a CBS tree prior}
\usage{
remove_cbs_exps(experiments)
}
\description{
Remove all experiments that have a CBS tree prior
}
