% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_razzo.R
\name{run_razzo}
\alias{run_razzo}
\title{Runs one \code{razzo} experiment}
\usage{
run_razzo(razzo_params)
}
\arguments{
\item{razzo_params}{a parameter set for one \code{razzo} experiment,
as created by \link{create_params_razzo}
and used by \link{run_razzo}}
}
\value{
a data frame with errors,
  with as many rows as model selection parameter sets.
  Tip: use \link[pirouette]{pir_plot} to display it.
  More important are the files it creates.
}
\description{
Runs one \code{razzo} experiment
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
