% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_razzo_from_file.R
\name{run_razzo_from_file}
\alias{run_razzo_from_file}
\title{Run a razzo experiment from a filename}
\usage{
run_razzo_from_file(parameters_filename)
}
\arguments{
\item{parameters_filename}{name of a file that contains
a \code{razzo_params},
as can be created by \link{create_razzo_params}}
}
\description{
Run a razzo experiment from a filename
}
\author{
Richel J.C. Bilderbeek, Giovanni Laudanno
}
