% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_files_razzo_paramses.R
\name{save_razzo_paramses}
\alias{save_razzo_paramses}
\title{Create the parameter files according to the specified arguments
  \code{project_folder_name/data/[settings]/seed/[models]}}
\usage{
save_razzo_paramses(project_folder_name,
  mbd_paramses = create_mbd_params_table(lambda = 0.2, mu = 0.15, nu =
  c(1, 1.5, 2), q = c(0.1, 0.15, 0.2), n_replicates = 2, crown_age = 15,
  cond = 1),
  twinning_params = pirouette::create_twinning_params(twin_tree_filename
  = peregrine::get_pff_tempfile(), twin_alignment_filename =
  peregrine::get_pff_tempfile(), twin_evidence_filename =
  peregrine::get_pff_tempfile()),
  alignment_params = pirouette::create_alignment_params(root_sequence =
  pirouette::create_blocked_dna(length = 1000), mutation_rate =
  0.5/unique(mbd_paramses$crown_age), fasta_filename =
  peregrine::get_pff_tempfile(pattern = "alignment_", fileext = ".fasta")),
  error_measure_params = pirouette::create_error_measure_params(),
  mcmc_chain_length = beautier::create_mcmc()$chain_length)
}
\arguments{
\item{project_folder_name}{project folder name,
will be the full path to \code{razzo_project}}

\item{mbd_paramses}{data frame with one MBD parameter set per row.
If a simulation with one MBD parameter set has 10 replicates,
this data frame will hold 10 rows, each with different seeds}

\item{twinning_params}{parameters for creating a twin tree,
as can be created by \code{\link[pirouette]{create_twinning_params}}}

\item{alignment_params}{parameters for creating an alignment,
as can be created by \code{\link[pirouette]{create_alignment_params}}}

\item{error_measure_params}{parameters to set how the error
between given tree and inferred trees in measure,
as can be created by \code{\link[pirouette]{create_error_measure_params}}}

\item{mcmc_chain_length}{length of an MCMC}
}
\value{
Create folders for each parameter setting
  and saves each setting in a file within the corresponding folder.
}
\description{
Create the parameter files according to the specified arguments
  \code{project_folder_name/data/[settings]/seed/[models]}
}
\author{
Giovanni Laudanno, Richel J.C. Bilderbeek
}
